### 2024-09-03 #7171 @editorialbot
[PRE REVIEW]: lintr: Static Code Analysis for R
**Submitting author:** @jimhester (<a href="http://orcid.org/0000-0002-2739-7082">James Hester</a>)
**Repository:** https://github.com/r-lib/lintr
**Branch with paper.md** (empty if default branch):
**Version:** v3.1.2
