+++ paper/paper.Rmd
@@ -38,7 +38,7 @@ affiliations:
   - index: 6
     name: The University of Queensland
   - index: 7
-    name: ???
+    name: Unknown
 output: 
     md_document:
       variant: "markdown"
