### 2024-09-05 136a61b
Remove a few linters from efficiency tag
closes #2653
