### 2024-09-05 752ba12
Update NEWS.md
