+++ NEWS.md
@@ -18,7 +18,7 @@
 * `cyclocomp_linter()` is no longer part of the default linters (#2555, @IndrajeetPatil) because the tidyverse style guide doesn't contain any guidelines on meeting certain complexity requirements. Note that users with `cyclocomp_linter()` in their configs may now need to install {cyclocomp} intentionally, in particular in CI/CD pipelines.
 * `scalar_in_linter` is now configurable to allow other `%in%` like operators to be linted. The data.table operator `%chin%` is no longer linted by default; use `in_operators = "%chin%"` to continue linting it. (@F-Noelle)
 * `lint()` and friends now normalize paths to forward slashes on Windows (@olivroy, #2613).
-* Following linters have been removed from efficiency linters (@IndrajeetPatil):
+* Following linters have been removed from efficiency linters (@IndrajeetPatil, #2655):
   `undesirable_function_linter()`, `undesirable_operator_linter()`, and `list_comparison_linter()`
   
 
