### 2024-09-05 ba6af1c
Find some = or -> assignments when discovering script globals
