### 2024-09-05 491c339
Update NEWS.md
Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>
