### 2024-09-05 7fa4876
Remove a few linters from efficiency tag (#2655)
* Remove a few linters from efficiency tag

closes #2653

* Update NEWS.md

* Update NEWS.md

Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>

---------

Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>
