### 2024-09-06 427fab3
Find some = or -> assignments when discovering script globals (#2656)
Co-authored-by: AshesITR <alexander.rosenstock@web.de>
