### 2024-09-06 #7171 @jimhester
@IndrajeetPatil You may want to look and see if we can resolve the issues flagged in this thread, particularly the license and author information above (though I do not know Florent Angly's affiliation)
