### 2024-09-06 #7171 @MichaelChirico
thanks for flagging @jimhester! it was (partially) fixed already: https://github.com/r-lib/lintr/pull/2651
