### 2024-09-06 562534e
Leave out `sort_linter()` from efficiency linters
https://github.com/r-lib/lintr/issues/2653#issuecomment-2333476336
