+++ man/efficiency_linters.Rd
@@ -34,7 +34,6 @@ The following linters are tagged with 'efficiency':
 \item{\code{\link{routine_registration_linter}}}
 \item{\code{\link{sample_int_linter}}}
 \item{\code{\link{scalar_in_linter}}}
-\item{\code{\link{seq_linter}}}
 \item{\code{\link{sort_linter}}}
 \item{\code{\link{string_boundary_linter}}}
 \item{\code{\link{unnecessary_concatenation_linter}}}
