### 2024-09-06 7c947ce
Include `{data.table}` example in `redundant_equals_linter()` docs
closes #2657
