+++ R/redundant_equals_linter.R
@@ -19,6 +19,11 @@
 #'   linters = redundant_equals_linter()
 #' )
 #'
+#' lint(
+#'   text = "dt[x == FALSE, y]",
+#'   linters = redundant_equals_linter()
+#' )
+#'
 #' # okay
 #' lint(
 #'   text = "if (any(x)) 1",
@@ -27,6 +32,11 @@
 #'
 #' lint(
 #'   text = "if (!all(x)) 0",
+#'   linters = redundant_equals_linter()
+#' )
+#'
+#' lint(
+#'   text = "dt[!(x), y]",
 #'   linters = redundant_equals_linter()
 #' )
 #'
