### 2024-09-07 25aa6f9
Update R/redundant_equals_linter.R
Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>
