+++ R/redundant_equals_linter.R
@@ -20,7 +20,7 @@
 #' )
 #'
 #' lint(
-#'   text = "dt[x == FALSE, y]",
+#'   text = "dt[is_tall == FALSE, y]",
 #'   linters = redundant_equals_linter()
 #' )
 #'
@@ -35,9 +35,9 @@
 #'   linters = redundant_equals_linter()
 #' )
 #'
-#' # in data.table semantics, dt[x] is a join, dt[(x)] is a subset
+#' # in `{data.table}` semantics, `dt[x]` is a join, `dt[(x)]` is a subset
 #' lint(
-#'   text = "dt[!(x), y]",
+#'   text = "dt[(!is_tall), y]",
 #'   linters = redundant_equals_linter()
 #' )
 #'
