### 2024-09-07 9ff8e37
simplify test as suggested in review
