+++ tests/testthat/test-linter_tags.R
@@ -236,9 +236,5 @@ test_that("available_linters gives precedence to included tags", {
 })
 
 test_that("all non-deprecated linters are covered across tags", {
-  tags <- setdiff(available_tags(), "deprecated")
-  tagged_linters <- lapply(tags, function(tag) names(linters_with_tags(tag))) # nolint: undesirable_function_linter.
-  tagged_linters <- unique(unlist(tagged_linters))
-  all_linters <- names(all_linters())
-  expect_identical(sort(tagged_linters), sort(all_linters))
+  expect_true(all(lengths(available_linters()$tags) > 0L))
 })
