### 2024-09-09 #7171 @lrnv
@jimhester Happy to be your editor for this meaningful package in the R realm.

To start off, would you be able to recommend a few potential reviewers for this ? For example is you have people that you know will give you interesting input on the submission, that might fit.
