### 2024-09-09 #7171 @MichaelChirico
Also looked briefly; maybe https://reviewers.joss.theoj.org/reviewers/2436 who has worked on the R Dev Guide & other R forwards initiatives
