### 2024-09-10 #7171 @lrnv
Okay then, @njtierney, @josiahparry, @SaranjeetKaur would any of you be willing to review `r-lib/lintr` for JOSS ? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
