### 2024-09-16 #7240 @lrnv
@JosiahParry, @SaranjeetKaur greetings! Thanks for accepting to take a bit of time te review this submission. FIrst, do you know how JOSS review are handled, or do you need me to wrap it up for you ? The first thing you have to do is to generate a guide for you -- formatted as a checklist --, using a command `@editorialbot generate my checklist.` in this discussion. Then if you have more questions, I'll be happy to help
