### 2024-09-16 #7240 @JosiahParry
The package, documentation, and testing is great. Please address the below regarding the paper. Notably there isn't a clear summary of lintr for a non-specialist audience, nor is there a state of the field section.

## Paper

- Please define what a linter is and what their utility is
- please make the summary of lintr more explicit than
> "The `{lintr}` package is an open-source R package that provides static code analysis to check for a variety of common problems related to readability, efficiency, consistency, style, etc."
- Please provide a description of other static code analysis packages—e.g. luke tierneys codetools, Mango's goodpractice, and more recently flint
- Please provide a description of what a linter is prior discussing some of the 113 provided by {lintr}.

- Under **Efficiency** the use of "the users" can remove `"the"`.
```diff
- Sometimes the users might not be aware
+ Sometimes users might not be aware
```
