+++ paper/paper.bib
@@ -32,10 +32,44 @@
     publisher={Pearson Education}
   }
 
-  @misc{ enwiki:1218663830,
+@misc{ enwiki:1218663830,
     author = "{Wikipedia contributors}",
     title = "Static program analysis --- {Wikipedia}{,} The Free Encyclopedia",
     year = "2024",
     url = "https://en.wikipedia.org/w/index.php?title=Static_program_analysis&oldid=1218663830",
     note = "[Online; accessed 7-May-2024]"
+  }
+
+@Manual{Tierney2024,
+    title = {codetools: Code Analysis Tools for R},
+    author = {Luke Tierney},
+    year = {2024},
+    note = {R package version 0.2-20},
+    url = {https://CRAN.R-project.org/package=codetools},
+  }
+
+@Manual{Bacher2024,
+    title = {flint: Find and Fix Lints in R Code},
+    author = {Etienne Bacher},
+    year = {2024},
+    note = {R package version 0.0.2,
+https://github.com/etiennebacher/flint},
+    url = {https://flint.etiennebacher.com},
+  }
+
+@Manual{Müller2024,
+    title = {styler: Non-Invasive Pretty Printing of R Code},
+    author = {Kirill Müller and Lorenz Walthert and Indrajeet Patil},
+    year = {2024},
+    note = {R package version 1.10.3.9000, commit 6d2f0b34245b6bc712bf2fcabf240d9ca814f0ef},
+    url = {https://github.com/r-lib/styler},
+  }
+
+  @Manual{Padgham2024,
+    title = {goodpractice: Advice on R Package Building},
+    author = {Mark Padgham and Karina Marks and Daniel {de Bortoli} and Gabor Csardi and Hannah Frick and Owen Jones and Hannah Alexander},
+    year = {2024},
+    note = {R package version 1.0.5,
+https://github.com/ropensci-review-tools/goodpractice},
+    url = {https://docs.ropensci.org/goodpractice/},
   }
