### 2024-09-25 09fb9da
Update paper/paper.Rmd
Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>
