+++ paper/paper.bib
@@ -49,7 +49,7 @@
   }
 
 @Manual{Bacher2024,
-    title = {flint: Find and Fix Lints in R Code},
+    title = {flint: Find and Fix Lints in {R} Code},
     author = {Etienne Bacher},
     year = {2024},
     note = {R package version 0.0.2,
@@ -65,11 +65,43 @@ https://github.com/etiennebacher/flint},
     url = {https://github.com/r-lib/styler},
   }
 
-  @Manual{Padgham2024,
+@Manual{Padgham2024,
     title = {goodpractice: Advice on R Package Building},
     author = {Mark Padgham and Karina Marks and Daniel {de Bortoli} and Gabor Csardi and Hannah Frick and Owen Jones and Hannah Alexander},
     year = {2024},
     note = {R package version 1.0.5,
 https://github.com/ropensci-review-tools/goodpractice},
     url = {https://docs.ropensci.org/goodpractice/},
+  }
+
+@Manual{Rudolph2024,
+    title = {box: Write Reusable, Composable and Modular {R} Code},
+    author = {Konrad Rudolph},
+    year = {2024},
+    note = {R package version 1.2.0},
+    url = {https://CRAN.R-project.org/package=box},
+  }
+
+@Manual{Basa2024,
+    title = {box.linters: Linters for 'box' Modules},
+    author = {Ricardo Rodrigo Basa and Jakub Nowicki},
+    year = {2024},
+    note = {R package version 0.10.5},
+    url = {https://CRAN.R-project.org/package=box.linters},
+  }
+
+@Manual{Kelkhoff2024,
+    title = {roxylint: Lint 'roxygen2'-Generated Documentation},
+    author = {Doug Kelkhoff},
+    year = {2024},
+    note = {R package version 0.1.0},
+    url = {https://CRAN.R-project.org/package=roxylint},
+  }
+
+@Manual{Wickham2024roxy,
+    title = {roxygen2: In-Line Documentation for {R}},
+    author = {Hadley Wickham and Peter Danenberg and Gábor Csárdi and Manuel Eugster},
+    year = {2024},
+    note = {R package version 7.3.2},
+    url = {https://CRAN.R-project.org/package=roxygen2},
   }
