+++ paper/paper.Rmd
@@ -46,7 +46,7 @@ output:
       standalone: true
 bibliography: paper.bib
 csl: apa.csl
-link-citations: yes
+link-citations: true
 ---
 
 ```{r setup, warning=FALSE, message=FALSE, echo=FALSE}
@@ -69,7 +69,7 @@ A linter is a tool that analyzes code to identify potential errors, stylistic is
 
 # Features
 
-By default, `{lintr}` enforces the tidyverse style guide [@Wickham2023,@Müller2024]. In this respect, it differs from other static code analysis tools in R (like `{codetools}` [@Tierney2024]), which are not opinionated and don't enforce any particular style of writing code, but, rather, check R code for possible problems.
+By default, `{lintr}` enforces the tidyverse style guide [@Wickham2023,@Müller2024]. In this respect, it differs from other static code analysis tools in R (like `{codetools}` [@Tierney2024]), which are not opinionated and don't enforce any particular style of writing code, but, rather, check R code for possible problems. Additionally, `{lintr}` doesn't cover linting inline code for documentation generation (cf. `{roxylint}` [@Kelkhoff2024] for linting `{roxygen2}` comments [@Wickham2024roxy]).
 
 As of this writing, `{lintr}` offers `r length(all_linters())` linters.
 
@@ -219,7 +219,7 @@ There are two main ways to customize it:
 
   - Create new linters (by leveraging functions like `lintr::make_linter_from_xpath()`) tailored to match project- or organization-specific coding standards.
 
-Indeed, `{goodpractice}` [@Padgham2024] bundles a set of custom linters that are not part of the default set of `{lintr}` linters, while `{flint}` [@Bacher2024] provides a way to fix the lints detected by `{lintr}`.
+Indeed, `{goodpractice}` [@Padgham2024] bundles a set of custom linters that are not part of the default set of `{lintr}` linters, while `{flint}` [@Bacher2024] provides a way to fix the lints detected by `{lintr}`. Moreover, `{box.linters}` [@Basa2024] extends `{lintr}` to support `{box}` modules [@Rudolph2024].
 
 # Benefits of using `{lintr}`
 
