+++ paper/paper.bib
@@ -105,3 +105,13 @@ https://github.com/ropensci-review-tools/goodpractice},
     note = {R package version 7.3.2},
     url = {https://CRAN.R-project.org/package=roxygen2},
   }
+
+@Manual{Onkelinx2024,
+    title = {checklist: A Thorough and Strict Set of Checks for {R} Packages and Source Code. Version 0.4.0},
+    author = {Thierry Onkelinx},
+    year = {2024},
+    url = {https://inbo.github.io/checklist/},
+    abstract = {An opinionated set of rules for R packages and R source code projects.},
+    keywords = {quality control; documentation; publication},
+    doi = {10.5281/zenodo.4028303},
+  }
