### 2024-10-02 0c2c3f6
Update paper.Rmd
Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>
