+++ paper/paper.md
@@ -1,6 +1,6 @@
 ---
 title: "Static Code Analysis for R"
-date: "2024-09-05"
+date: "2024-10-02"
 tags: ["R", "linter", "tidyverse"]
 authors:
   - name: Jim Hester
@@ -46,26 +46,36 @@ output:
       standalone: true
 bibliography: paper.bib
 csl: apa.csl
-link-citations: yes
+link-citations: true
 ---
 
 # Statement of Need
 
-R is an interpreted, dynamically-typed programming language [@base2023].
-It is a popular choice for statistical analysis and visualization, and
-is used by a wide range of researchers and data scientists. The
-`{lintr}` package is an open-source R package that provides static code
-analysis [@enwiki:1218663830] to check for a variety of common problems
-related to readability, efficiency, consistency, style, etc. In
-particular, by default it enforces the tidyverse style guide
-[@Wickham2023]. It is designed to be easy to use and integrate into
-existing workflows, and can be used as part of an automated build or
-continuous integration process. `{lintr}` also integrates with a number
-of popular IDEs and text editors, such as RStudio and Visual Studio
-Code, making it convenient for users to run `{lintr}` checks on their
-code as they work.
+A linter is a tool that analyzes code to identify potential errors,
+stylistic issues, or deviations from coding standards. It helps ensure
+consistency, readability, and best practices by flagging common
+mistakes, such as syntax errors, unused variables, or improper
+formatting. Linters are essential for improving code quality, preventing
+bugs, and maintaining a clean codebase, especially in collaborative
+development environments [@enwiki:1218663830]. `{lintr}` is an
+open-source package that provides linters for the R programming
+language, which is an interpreted, dynamically-typed programming
+language [@base2023], and is used by a wide range of researchers and
+data scientists. `{lintr}` can thus act as a valuable tool for R users
+to help improve the quality and reliability of their code.
 
 # Features
+
+By default, `{lintr}` enforces the tidyverse style guide
+[@Wickham2023,@Müller2024]. In this respect, it differs from other
+static code analysis tools in R (like `{codetools}` [@Tierney2024]),
+which are not opinionated and don't enforce any particular style of
+writing code, but, rather, check R code for possible problems
+(incidentally, `{lintr}` uses `{codetools}` as a backend for object
+usage linters). Additionally, `{lintr}` is concerned only with R code,
+so code-adjacent text like inline `{roxygen2}` comments
+[@Wickham2024roxy] will not be covered (cf. `{roxylint}`
+[@Kelkhoff2024]).
 
 As of this writing, `{lintr}` offers 113 linters.
 
@@ -115,9 +125,9 @@ lint(
 
 -   **Efficiency**
 
-Sometimes the users might not be aware of a more efficient way offered
-by R for carrying out a computation. `{lintr}` offers linters to improve
-code efficiency by avoiding common inefficient patterns.
+Sometimes users might not be aware of a more efficient way offered by R
+for carrying out a computation. `{lintr}` offers linters to improve code
+efficiency by avoiding common inefficient patterns.
 
 For example, the `any_is_na_linter()` linter detects usages of
 `any(is.na(x))` and suggests `anyNA(x)` as a more efficient alternative
@@ -285,6 +295,14 @@ lint(
     `lintr::make_linter_from_xpath()`) tailored to match project- or
     organization-specific coding standards.
 
+Indeed, `{goodpractice}` [@Padgham2024] bundles a set of custom linters
+that are not part of the default set of `{lintr}` linters, while
+`{box.linters}` [@Basa2024] extends `{lintr}` to support `{box}` modules
+[@Rudolph2024] and `{checklist}` includes linters as one of the strict
[…]
