### 2024-10-02 #7240 @IndrajeetPatil
@JosiahParry Thanks a lot for your feedback! We have updated the draft to address your concerns.
Please let us know if you have any additional feedback!

---

P.S. We are also currently looking into a possible encoding issue seen in the PDF output:
