### 2024-10-02 #7240 @lrnv
@IndrajeetPatil Happens sometimes in Julia when some characters are not availiable in the monospace font used. You can change the default monospace font (for one that includes the glyph you need) by adding (chnaging the font):

````
header-includes:
- |
  ```{=latex}
  \setmonofont[Path=./]{JuliaMono-Regular.ttf}
  ```
````

to the yaml header of the paper. See https://github.com/openjournals/joss/issues/963 for wide discussions if this does not work.
