### 2024-10-03 0d55aa0
Fix missing character issue in JOSS's PDF output
