+++ man/style_linters.Rd
@@ -16,6 +16,7 @@ The following linters are tagged with 'style':
 \item{\code{\link{brace_linter}}}
 \item{\code{\link{commas_linter}}}
 \item{\code{\link{commented_code_linter}}}
+\item{\code{\link{complex_conditional_linter}}}
 \item{\code{\link{condition_call_linter}}}
 \item{\code{\link{consecutive_assertion_linter}}}
 \item{\code{\link{cyclocomp_linter}}}
