+++ man/complex_conditional_linter.Rd
@@ -1,0 +1,40 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/complex_conditional_linter.R
+\name{complex_conditional_linter}
+\alias{complex_conditional_linter}
+\title{Complex Conditional Expressions Linter}
+\usage{
+complex_conditional_linter(threshold = 2L)
+}
+\arguments{
+\item{threshold}{Integer. The maximum number of logical operands (\code{&&} or \code{||})
+allowed in a conditional expression (default: \code{2L}).}
+}
+\description{
+Detects complex conditional expressions and suggests extracting
+them into Boolean functions or variables for improved readability and reusability.
+}
+\examples{
+# This will produce a lint because there are three logical operands
+code <- "if (a && b && c) { do_something() }"
+writeLines(code)
+lint(
+  text = code,
+  linters = complex_conditional_linter()
+)
+
+# This will not produce a lint because it meets the threshold
+code <- "if (a && b) { do_something() }"
+writeLines(code)
+lint(
+  text = code,
+  linters = complex_conditional_linter()
+)
+
+}
+\seealso{
+\link{linters} for a complete list of linters available in lintr.
+}
+\section{Tags}{
+\link[=best_practices_linters]{best_practices}, \link[=configurable_linters]{configurable}, \link[=readability_linters]{readability}, \link[=style_linters]{style}
+}
