+++ man/configurable_linters.Rd
@@ -17,6 +17,7 @@ The following linters are tagged with 'configurable':
 \item{\code{\link{backport_linter}}}
 \item{\code{\link{brace_linter}}}
 \item{\code{\link{commas_linter}}}
+\item{\code{\link{complex_conditional_linter}}}
 \item{\code{\link{condition_call_linter}}}
 \item{\code{\link{conjunct_test_linter}}}
 \item{\code{\link{consecutive_mutate_linter}}}
