+++ NAMESPACE
@@ -31,6 +31,7 @@ export(closed_curly_linter)
 export(commas_linter)
 export(commented_code_linter)
 export(comparison_negation_linter)
+export(complex_conditional_linter)
 export(condition_call_linter)
 export(condition_message_linter)
 export(conjunct_test_linter)
