### 2024-10-20 c838ecd
fix: remove base pipe from end of line before detecting parsability
