### 2024-10-20 fec6829
fix: remove magrittr pipe from end of line before detecting parsability
