+++ tests/testthat/test-commented_code_linter.R
@@ -103,3 +103,19 @@ test_that("commented_code_linter can detect operators in comments and lint corre
     commented_code_linter()
   )
 })
+
+test_that("commented_code_linter can detect commented code ending with a pipe", {
+  skip_if_not_r_version("4.1.0")
+
+  expect_lint(
+    "# f() |>",
+    rex::rex("Remove commented code."),
+    commented_code_linter()
+  )
+
+  expect_lint(
+    "# f() %>%",
+    rex::rex("Remove commented code."),
+    commented_code_linter()
+  )
+})
