### 2024-10-20 95b95db
chore: add commented code linter ignoring end pipe to news
