### 2024-10-22 cc0a066
fix: seperate tests for magrittr/base pipe to allow for minimum R version
