### 2024-10-22 0d4c655
refac: check for trailing comma/pipes in one fn call
