+++ R/commented_code_linter.R
@@ -83,12 +83,9 @@ commented_code_linter <- function() {
     all_comments <- xml_text(all_comment_nodes)
     code_candidates <- re_matches(all_comments, code_candidate_regex, global = FALSE, locations = TRUE)
     extracted_code <- code_candidates[, "code"]
-    # ignore trailing ',' when testing for parsability
-    extracted_code <- re_substitutes(extracted_code, rex(",", any_spaces, end), "")
+    # ignore trailing ',' or pipes ('|>', '%>%') when testing for parsability
+    extracted_code <- re_substitutes(extracted_code, rex(or(",", "|>", "%>%"), any_spaces, end), "")
     extracted_code <- re_substitutes(extracted_code, rex(start, any_spaces, ","), "")
-    # ignore a trailing pipe (|> or %>%) when testing parsability
-    extracted_code <- re_substitutes(extracted_code, rex("|>", any_spaces, end), "")
-    extracted_code <- re_substitutes(extracted_code, rex("%>%", any_spaces, end), "")
 
     is_parsable <- which(vapply(extracted_code, parsable, logical(1L)))
 
