### 2024-10-24 46325ab
chore: minor rewording of news
Co-authored-by: AshesITR <alexander.rosenstock@web.de>
