### 2024-10-24 b3e8657
further condense test code
