### 2024-10-26 78b0130
accept numeric values
