+++ man/complex_conditional_linter.Rd
@@ -51,8 +51,8 @@ if (is_duck) \{
 \}
 }\if{html}{\out{</div>}}
 
-In addition to improving code readability, extracting complex conditional expressions has the added benefit
-of introducing a reusable abstraction.
+In addition to improving code readability, extracting complex conditional expressions
+has the added benefit of introducing a reusable abstraction.
 }
 \examples{
 # will produce lints
@@ -64,12 +64,18 @@ lint(
 )
 
 # okay
-ready_to_do_something <- a && b && c
 code <- "if (ready_to_do_something) { do_something() }"
 writeLines(code)
 lint(
   text = code,
   linters = complex_conditional_linter()
+)
+
+code <- "if (a && b && c) { do_something() }"
+writeLines(code)
+lint(
+  text = code,
+  linters = complex_conditional_linter(threshold = 2L)
 )
 
 }
