### 2024-10-27 32a2229
re_matches_logical helper ensures logical output of re_matches_logical
