+++ tests/testthat/test-expect_lint.R
@@ -30,7 +30,7 @@ test_that("single check", {
   expect_success(expect_lint("a=1", list(message = lint_msg, line_number = 1L), linter))
   expect_failure(expect_lint("a=1", list(2L, lint_msg), linter))
 
-  expect_error(expect_lint("1:nrow(x)", "(group)", seq_linter()), "Invalid regex result", fixed = TRUE)
+  expect_success(expect_lint("1:nrow(x)", "(nrow)", seq_linter()))
 })
 
 test_that("multiple checks", {
