### 2024-10-30 fe71427
Test for new beahaviour in assigmnet_linter.r
New behaviour not yet implemented
