### 2024-11-01 746eba2
change default and check for input edge cases
