### 2024-11-01 bea5900
Update test-complex-conditional-linter.R
