+++ man/efficiency_linters.Rd
@@ -39,6 +39,7 @@ The following linters are tagged with 'efficiency':
 \item{\code{\link{string_boundary_linter}}}
 \item{\code{\link{unnecessary_concatenation_linter}}}
 \item{\code{\link{unnecessary_lambda_linter}}}
+\item{\code{\link{unnecessary_which_linter}}}
 \item{\code{\link{vector_logic_linter}}}
 \item{\code{\link{which_grepl_linter}}}
 }
