+++ man/unnecessary_which_linter.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/unnecessary_which_linter.R
+\name{unnecessary_which_linter}
+\alias{unnecessary_which_linter}
+\title{Find usage of \code{which(x == min(x))} and \code{which(x == max(x))}}
+\usage{
+unnecessary_which_linter()
+}
+\description{
+\code{which.min()} and \code{which.max()} are optimized functions that directly return
+the index of the minimum or maximum value in a vector. They are more efficient
+and cleaner than using \code{which()} with equality comparison to \code{min()} or \code{max()}.
+}
+\details{
+This linter identifies code using \code{which()} with equality comparison to \code{min()}
+or \code{max()}, suggesting the use of more efficient \code{which.min()} and \code{which.max()}
+functions instead.
+}
+\seealso{
+\link{linters} for a complete list of linters available in lintr.
+}
+\section{Tags}{
+\link[=best_practices_linters]{best_practices}, \link[=efficiency_linters]{efficiency}, \link[=readability_linters]{readability}
+}
