+++ man/linters.Rd
@@ -17,18 +17,18 @@ see also \code{\link[=available_tags]{available_tags()}}.
 \section{Tags}{
 The following tags exist:
 \itemize{
-\item{\link[=best_practices_linters]{best_practices} (63 linters)}
+\item{\link[=best_practices_linters]{best_practices} (64 linters)}
 \item{\link[=common_mistakes_linters]{common_mistakes} (11 linters)}
 \item{\link[=configurable_linters]{configurable} (44 linters)}
 \item{\link[=consistency_linters]{consistency} (32 linters)}
 \item{\link[=correctness_linters]{correctness} (7 linters)}
 \item{\link[=default_linters]{default} (25 linters)}
 \item{\link[=deprecated_linters]{deprecated} (6 linters)}
-\item{\link[=efficiency_linters]{efficiency} (29 linters)}
+\item{\link[=efficiency_linters]{efficiency} (30 linters)}
 \item{\link[=executing_linters]{executing} (6 linters)}
 \item{\link[=package_development_linters]{package_development} (14 linters)}
 \item{\link[=pkg_testthat_linters]{pkg_testthat} (12 linters)}
-\item{\link[=readability_linters]{readability} (64 linters)}
+\item{\link[=readability_linters]{readability} (65 linters)}
 \item{\link[=regex_linters]{regex} (4 linters)}
 \item{\link[=robustness_linters]{robustness} (17 linters)}
 \item{\link[=style_linters]{style} (40 linters)}
@@ -145,6 +145,7 @@ The following linters exist:
 \item{\code{\link{unnecessary_lambda_linter}} (tags: best_practices, configurable, efficiency, readability)}
 \item{\code{\link{unnecessary_nesting_linter}} (tags: best_practices, configurable, consistency, readability)}
 \item{\code{\link{unnecessary_placeholder_linter}} (tags: best_practices, readability)}
+\item{\code{\link{unnecessary_which_linter}} (tags: best_practices, efficiency, readability)}
 \item{\code{\link{unreachable_code_linter}} (tags: best_practices, configurable, readability)}
 \item{\code{\link{unused_import_linter}} (tags: best_practices, common_mistakes, configurable, executing)}
 \item{\code{\link{vector_logic_linter}} (tags: best_practices, common_mistakes, default, efficiency)}
