+++ NAMESPACE
@@ -154,6 +154,7 @@ export(unnecessary_lambda_linter)
 export(unnecessary_nested_if_linter)
 export(unnecessary_nesting_linter)
 export(unnecessary_placeholder_linter)
+export(unnecessary_which_linter)
 export(unneeded_concatenation_linter)
 export(unreachable_code_linter)
 export(unused_import_linter)
