### 2024-11-03 4e4d41c
add tests and examples
