### 2024-11-03 #7240 @SaranjeetKaur
Thanks for all your work! Great package and documentation!

### Paper

> Naturally, we can’t discuss all of them here. To see details about all available linters, we encourage readers to see https://lintr.r- lib.org/dev/reference/index.html#individual- linters.

This could perhaps be rephrased as (instead of putting a link, it could be a hyperlink):

_To see the most up-to-date details about all the available linters, we encourage readers to refer to the [list of individual linters](https://lintr.r- lib.org/dev/reference/index.html#individual- linters)._
