### 2024-11-03 #7240 @IndrajeetPatil
Happy to rephrase that way.

> instead of putting a link, it could be a hyperlink?

That's not friendly to readers who like to read the printed versions of publications (me being one of them 😉)
