### 2024-11-03 #7240 @SaranjeetKaur
> That's not friendly to readers who like to read the printed versions of publications (me being one of them 😉)

I see, I didn't realise that! Thanks for clarifying!
