### 2024-11-03 #7240 @SaranjeetKaur
Although the "Statement of Need" section is giving the required info, it talks about the tool before mentioning why the tool is required. So perhaps a sentence or two, might help clarify?

For example,

_In computer programming, "linting" is the process of analysing the source code to identify possible programming and stylistic problems (Refer: https://en.wiktionary.org/wiki/linting). This can be done using a software tool called a linter (or a lint tool). A linter analyzes code to identify potential errors, stylistic issues, or deviations from coding standards. This helps ensure consistency, readability, and best practices by flagging common mistakes, such as syntax errors, unused variables, or improper formatting. Linters are essential for improving code quality, preventing bugs, and maintaining a clean codebase, especially in collaborative development environments (Wikipedia contributors, 2024). {lintr} is an open-source package that provides linters for the R programming language, which is an interpreted, dynamically-typed programming language (R Core Team, 2023), and is used by a wide range of researchers and data scientists. {lintr} can thus act as a valuable tool for R users to help improve the quality and reliability of their code._
