### 2024-11-03 #7240 @SaranjeetKaur
Installing the development version from GitHub (`packageVersion("lintr") # ‘3.1.2.9000’`) gives length `113`:

```r
remotes::install_github("r-lib/lintr")

library(lintr)

length(all_linters())
#> [1] 113
```

Whereas installing the stable version on CRAN (`packageVersion("lintr") # ‘3.1.2’`) gives length `96`:

```r
install.packages("lintr")

library(lintr)

length(all_linters())
#> [1] 96
```

Do you think this should be explicitly mentioned, with say, _As of this writing, the development version of {lintr} from GitHub offers `113` linters_? Followed by,

```r
# install.packages("remotes")
remotes::install_github("r-lib/lintr")
library(lintr)

length(all_linters())

#> [1] 113
```
