### 2024-11-03 #7240 @SaranjeetKaur
I can't seem to find the contributing guidelines (for the community) in the package directory. Please let me know if I missed something - I was looking for something like the [dplyr's CONTRIBUTING.md](https://github.com/tidyverse/dplyr/blob/main/.github/CONTRIBUTING.md)
