### 2024-11-03 #7240 @IndrajeetPatil
We have two different guidelines:

- For those wishing to contribute: https://github.com/r-lib/lintr/blob/main/.github/CODE_OF_CONDUCT.md
- For those seeking help: https://github.com/r-lib/lintr/blob/main/.github/SUPPORT.md

The closest thing we have to contributing guidelines is a vignette explaining how to add new linters to the package:
https://lintr.r-lib.org/articles/creating_linters.html
