### 2024-11-03 #7240 @IndrajeetPatil
> Do you think this should be explicitly mentioned, with say, As of this writing, the development version of {lintr} from GitHub offers 113 linters? Followed by,

I am not sure. We might have a new release before this submission is accepted and published (cf. https://github.com/r-lib/lintr/issues/2392).
WDYT, @MichaelChirico?
