### 2024-11-03 #7240 @SaranjeetKaur
> We have two different guidelines:
>
> * For those wishing to contribute: https://github.com/r-lib/lintr/blob/main/.github/CODE_OF_CONDUCT.md
> * For those seeking help: https://github.com/r-lib/lintr/blob/main/.github/SUPPORT.md
>
> The closest thing we have to contributing guidelines is a vignette explaining how to add new linters to the package: https://lintr.r-lib.org/articles/creating_linters.html

Nice! I can see that the vignette on `creating_linters` talks about [submitting a pull request](https://lintr.r-lib.org/articles/creating_linters.html#submit-pull-request). Do you think it is worthwhile to create a `.github/CONTRIBUTING.md` file and appropriately link such stuff? It might potentially have sections like:

- [Fixing typos](https://github.com/tidyverse/dplyr/blob/main/.github/CONTRIBUTING.md#fixing-typos)
- [Pull request process](https://github.com/tidyverse/dplyr/blob/main/.github/CONTRIBUTING.md#pull-request-process)
- [Creating new linters](https://lintr.r-lib.org/articles/creating_linters.html): This section could provide a high level overview of what is already in the vignette.
