### 2024-11-04 #7240 @MichaelChirico
> We might have a new release before this submission is accepted and published

We can push for that if it's a priority, what's the timeline for when this paper would be published?
