### 2024-11-05 #7240 @lrnv
No timeline, it would be published when it is ready, but yes usually people synchronize that with a (at least minor) release, but this is not mandatory.
