+++ man/which_linter.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/which_grepl_linter.R
-\name{which_grepl_linter}
-\alias{which_grepl_linter}
+% Please edit documentation in R/which_linter.R
+\name{which_linter}
+\alias{which_linter}
 \title{Require usage of grep over which(grepl(.))}
 \usage{
-which_grepl_linter()
+which_linter()
 }
 \description{
 \code{which(grepl(pattern, x))} is the same as \code{grep(pattern, x)}, but harder
@@ -14,13 +14,13 @@ to read and requires two passes over the vector.
 # will produce lints
 lint(
   text = "which(grepl('^a', x))",
-  linters = which_grepl_linter()
+  linters = which_linter()
 )
 
 # okay
 lint(
   text = "which(grepl('^a', x) | grepl('^b', x))",
-  linters = which_grepl_linter()
+  linters = which_linter()
 )
 
 }
