+++ man/consistency_linters.Rd
@@ -43,7 +43,7 @@ The following linters are tagged with 'consistency':
 \item{\code{\link{system_file_linter}}}
 \item{\code{\link{T_and_F_symbol_linter}}}
 \item{\code{\link{unnecessary_nesting_linter}}}
-\item{\code{\link{which_grepl_linter}}}
+\item{\code{\link{which_linter}}}
 \item{\code{\link{whitespace_linter}}}
 }
 }
