+++ NAMESPACE
@@ -160,7 +160,7 @@ export(unreachable_code_linter)
 export(unused_import_linter)
 export(use_lintr)
 export(vector_logic_linter)
-export(which_grepl_linter)
+export(which_linter)
 export(whitespace_linter)
 export(with_defaults)
 export(with_id)
