+++ DESCRIPTION
@@ -198,7 +198,7 @@ Collate:
     'unused_import_linter.R'
     'use_lintr.R'
     'vector_logic_linter.R'
-    'which_grepl_linter.R'
+    'which_linter.R'
     'whitespace_linter.R'
     'with.R'
     'with_id.R'
