### 2024-11-05 40fcabd
Update NEWS.md
