+++ R/which_linter.R
@@ -13,12 +13,12 @@
 #' # will produce lints
 #' lint(
 #'   text = "which(x == max(x))",
-#'   linters = unnecessary_which_linter()
+#'   linters = which_linter()
 #' )
 #'
 #' lint(
 #'   text = "which(x == min(x))",
-#'   linters = unnecessary_which_linter()
+#'   linters = which_linter()
 #' )
 #'
 #' lint(
@@ -29,12 +29,12 @@
 #' # okay
 #' lint(
 #'   text = "which.max(x))",
-#'   linters = unnecessary_which_linter()
+#'   linters = which_linter()
 #' )
 #'
 #' lint(
 #'   text = "which.min(x))",
-#'   linters = unnecessary_which_linter()
+#'   linters = which_linter()
 #' )
 #'
 #' lint(
