### 2024-11-18 9d58027
Add utils to check for lint and error classes (#2681)
