### 2024-11-18 bea7bb1
Deploying to gh-pages from @ r-lib/lintr@9d58027882db0dcced8fbb2fd3d583ca0cc6315c 🚀
