### 2024-11-21 f6a5b61
base and internal copies not the same
