+++ R/utils.R
@@ -1,3 +1,12 @@
+`%||%` <- function(x, y) {
+  is_atomic_and_missing <- is.atomic(x[[1L]]) && is.na(x[[1L]])
+  if (is.null(x) || length(x) == 0L || is_atomic_and_missing) {
+    y
+  } else {
+    x
+  }
+}
+
 `%==%` <- function(x, y) {
   identical(x, y)
 }
