### 2024-11-29 #7240 @IndrajeetPatil
Thanks for the nudge, @lrnv. I have an open PR in the repo to address @SaranjeetKaur's feedback (https://github.com/r-lib/lintr/pull/2683). :)
