### 2024-11-29 #7240 @SaranjeetKaur
Thanks for the prompt response @IndrajeetPatil! The PR looks great!

Do you expect any discussion about the new release or mentioning the specific version of the package in the paper?

Otherwise, this looks all good to me. Thanks a lot for all the amazing work you all have put in!
