### 2024-11-29 #7240 @IndrajeetPatil
@SaranjeetKaur Yes, I would personally prefer to wait for this paper to be approved/published until the next version of this package is out on CRAN.

But I need to hand the baton over to @MichaelChirico at this point as he is the current maintainer.
