### 2024-12-01 #7240 @SaranjeetKaur
@IndrajeetPatil - it is all fine by me! I don't have any further comments. Could you or @MichaelChirico, please let us know whenever the next version of the package is out on CRAN?

@lrnv - is it okay if we have to wait for this paper to published until the next version of the package is out?
