### 2024-12-02 #7240 @MichaelChirico
Is {lintr} release the only current blocker, or are there other pending items? We are also trying to get {data.table} salvaged from the CRAN hammer so I need to prioritize OSS time :)
