### 2024-12-02 dae5c6f
Deploying to gh-pages from @ r-lib/lintr@57a9778f897529de2acde3592325df8e1fa5b281 🚀
