### 2024-12-02 b8195c2
Add a note about avoiding 'main'
