### 2024-12-02 b7cd0f6
Fix missing character issue in JOSS's PDF output (#2669)
