### 2024-12-02 f515956
Update .github/CONTRIBUTING.md
Co-authored-by: Michael Chirico <chiricom@google.com>
