+++ .github/CONTRIBUTING.md
@@ -32,7 +32,7 @@ If you wish to contribute a new linter, the [Creating new linters](https://lintr
     The title of your PR should briefly describe the change.
     The body of your PR should contain `Fixes #issue-number`.
 
-*  For user-facing changes, add a bullet to the top of `NEWS.md` (i.e. just below the first header). Follow the style described in <https://style.tidyverse.org/news.html>.
+*  For user-facing changes, add a bullet appropriately in the top section of `NEWS.md` (i.e. below the first header). Follow the style described in <https://style.tidyverse.org/news.html>.
 
 ### Code style
 
