### 2024-12-02 12b6c36
Update .github/CONTRIBUTING.md
Co-authored-by: Michael Chirico <chiricom@google.com>
