+++ .github/CONTRIBUTING.md
@@ -37,7 +37,7 @@ If you wish to contribute a new linter, the [Creating new linters](https://lintr
 ### Code style
 
 *   New code should follow the tidyverse [style guide](https://style.tidyverse.org). 
-    You can use the [styler](https://CRAN.R-project.org/package=styler) package to apply these styles, but please don't restyle code that has nothing to do with your PR.  
+    You can use the [styler](https://CRAN.R-project.org/package=styler) package to apply these styles.
 
 *  We use [roxygen2](https://cran.r-project.org/package=roxygen2), with [Markdown syntax](https://cran.r-project.org/web/packages/roxygen2/vignettes/rd-formatting.html), for documentation.  
 
