### 2024-12-02 ba61bb3
Remove unnecessary newlines in contributing.md
