### 2024-12-02 170df47
add cite to wiki page on linting
