+++ paper/paper.bib
@@ -32,12 +32,20 @@
     publisher={Pearson Education}
   }
 
-@misc{ enwiki:1218663830,
+@misc{enwiki:1218663830,
     author = "{Wikipedia contributors}",
     title = "Static program analysis --- {Wikipedia}{,} The Free Encyclopedia",
     year = "2024",
     url = "https://en.wikipedia.org/w/index.php?title=Static_program_analysis&oldid=1218663830",
-    note = "[Online; accessed 7-May-2024]"
+    note = "[Online; accessed 2-December-2024]"
+  }
+
+@misc{enwiki:1260589258,
+    author = "{Wikipedia contributors}",
+    title = "Lint (software) --- {Wikipedia}{,} The Free Encyclopedia",
+    year = "2024",
+    url = "https://en.wikipedia.org/w/index.php?title=Lint_(software)&oldid=1260589258",
+    note = "[Online; accessed 2-December-2024]"
   }
 
 @Manual{Tierney2024,
