+++ paper/paper.md
@@ -53,19 +53,19 @@ link-citations: true
 
 In computer programming, "linting" is the process of analyzing the
 source code to identify possible programming and stylistic problems
-(<https://en.wiktionary.org/wiki/linting>) and a linter is a tool used
-for linting. A linter analyzes code to identify potential errors,
-stylistic issues, or deviations from coding standards. It helps ensure
-consistency, readability, and best practices by flagging common
-mistakes, such as syntax errors, unused variables, or improper
-formatting. Linters are essential for improving code quality, preventing
-bugs, and maintaining a clean codebase, especially in collaborative
-development environments [@enwiki:1218663830]. `{lintr}` is an
-open-source package that provides linters for the R programming
-language, which is an interpreted, dynamically-typed programming
-language [@base2023], and is used by a wide range of researchers and
-data scientists. `{lintr}` can thus act as a valuable tool for R users
-to help improve the quality and reliability of their code.
+[@enwiki:1260589258] and a linter is a tool used for linting. A linter
+analyzes code to identify potential errors, stylistic issues, or
+deviations from coding standards. It helps ensure consistency,
+readability, and best practices by flagging common mistakes, such as
+syntax errors, unused variables, or improper formatting. Linters are
+essential for improving code quality, preventing bugs, and maintaining a
+clean codebase, especially in collaborative development environments
+[@enwiki:1218663830]. `{lintr}` is an open-source package that provides
+linters for the R programming language, which is an interpreted,
+dynamically-typed programming language [@base2023], and is used by a
+wide range of researchers and data scientists. `{lintr}` can thus act as
+a valuable tool for R users to help improve the quality and reliability
+of their code.
 
 # Features
 
