### 2024-12-02 6ad8769
Add note about extended Descriptions when needed, and hint at chains.
