### 2024-12-02 0eee980
clarify the audience of NEWS entries
