### 2024-12-02 bcb14f3
add NEWS item
