### 2024-12-02 0ec3122
Address feedback from reviewer-2 (#2683)
* Address feedback from reviewer-2

* Create CONTRIBUTING.md

* Add a note about avoiding 'main'

* Update .github/CONTRIBUTING.md

Co-authored-by: Michael Chirico <chiricom@google.com>

* Update .github/CONTRIBUTING.md

Co-authored-by: Michael Chirico <chiricom@google.com>

* Remove unnecessary newlines in contributing.md

* add cite to wiki page on linting

* emphasize why 'Fixes #' in description

* Add note about extended Descriptions when needed, and hint at chains.

* clarify the audience of NEWS entries

* tweak

---------

Co-authored-by: Michae
[…]
