### 2024-12-02 3f80509
Change in expect_lint() is also user-visible, so tweak NEWS
