### 2024-12-02 bce2a11
Deploying to gh-pages from @ r-lib/lintr@0ec3122686335bdada4012a8fcc5f46076201d73 🚀
