+++ R/lint.R
@@ -183,6 +183,7 @@ lint_dir <- function(path = ".", ...,
   }
 
   lints <- flatten_lints(lapply(
+    # NB: This cli API is experimental (https://github.com/r-lib/cli/issues/709)
     cli::cli_progress_along(files, name = "Running linters"),
     function(idx) lint(files[idx], ..., parse_settings = FALSE, exclusions = exclusions)
   ))
