### 2024-12-02 3f41984
respect `show_progress`'s authoritah
