+++ R/lint.R
@@ -182,12 +182,17 @@ lint_dir <- function(path = ".", ...,
     return(lints)
   }
 
-  lints <- flatten_lints(lapply(
-    # NB: This cli API is experimental (https://github.com/r-lib/cli/issues/709)
-    cli::cli_progress_along(files, name = "Running linters"),
-    function(idx) lint(files[idx], ..., parse_settings = FALSE, exclusions = exclusions)
-  ))
+  if (isTRUE(show_progress)) {
+    lints <- lapply(
+      # NB: This cli API is experimental (https://github.com/r-lib/cli/issues/709)
+      cli::cli_progress_along(files, name = "Running linters"),
+      function(idx) lint(files[idx], ..., parse_settings = FALSE, exclusions = exclusions)
+    )
+  } else {
+    lints <- lapply(files, lint, file, ..., parse_settings = FALSE, exclusions = exclusions)
+  }
 
+  lints <- flatten_lints(lints)
   lints <- reorder_lints(lints)
 
   if (relative_path) {
