### 2024-12-03 0103bad
re_matches_logical helper ensures logical output of re_matches_logical (#2679)
* re_matches_logical helper ensures logical output of re_matches_logical

* fix broken test

* Tests

* missed roxygenize()

* add NEWS item

* Change in expect_lint() is also user-visible, so tweak NEWS
