+++ .github/workflows/test-package-vigilant.yaml
@@ -28,7 +28,15 @@ jobs:
 
       - name: Run Tests
         run: |
-          options(crayon.enabled = TRUE)
+          ## --------------------------------------------------------------------
+          options(
+            crayon.enabled = TRUE,
+            warn = 2L,
+            warnPartialMatchArgs = TRUE,
+            warnPartialMatchAttr = TRUE,
+            warnPartialMatchDollar = TRUE
+          )
+          if (Sys.getenv("_R_CHECK_FORCE_SUGGESTS_", "") == "") Sys.setenv("_R_CHECK_FORCE_SUGGESTS_" = "false")
           pkgload::load_all()
           test_script_paths <- testthat::find_test_scripts("tests/testthat")
           test_with_warning_as_error <- function(path) {
