### 2024-12-03 eeb3215
Register Enhances S3 methods more selectively
