### 2024-12-03 13c7682
Use `{cli}` progress bar (#2641)
* Use cli progress bar: first draft

* update NEWS

* Update NEWS.md

* tweak

* note experimental status in a comment

* respect `show_progress`'s authoritah

* fix warning

* nolint

---------

Co-authored-by: Michael Chirico <chiricom@google.com>
