+++ NAMESPACE
@@ -181,13 +181,10 @@ importFrom(stats,na.omit)
 importFrom(tools,R_user_dir)
 importFrom(utils,capture.output)
 importFrom(utils,getParseData)
-importFrom(utils,getTxtProgressBar)
 importFrom(utils,globalVariables)
 importFrom(utils,head)
 importFrom(utils,relist)
-importFrom(utils,setTxtProgressBar)
 importFrom(utils,tail)
-importFrom(utils,txtProgressBar)
 importFrom(xml2,as_list)
 importFrom(xml2,xml_attr)
 importFrom(xml2,xml_children)
