### 2024-12-03 d8764fe
try without lambda
