### 2024-12-04 5a210e2
Use exportS3Method for suggest methods and move data.table to Suggests
