+++ NAMESPACE
@@ -2,6 +2,7 @@
 
 S3method("[",lints)
 S3method(as.data.frame,lints)
+S3method(data.table::as.data.table,lints)
 S3method(format,lint)
 S3method(format,lints)
 S3method(names,lints)
@@ -9,6 +10,7 @@ S3method(print,lint)
 S3method(print,lints)
 S3method(split,lints)
 S3method(summary,lints)
+S3method(tibble::as_tibble,lints)
 export(Lint)
 export(Linter)
 export(T_and_F_symbol_linter)
