### 2024-12-04 0de1d6d
Format with devel-`{styler}` (#2692)
* Format with devel-`{styler}`

* Update indentation_linter.R
