+++ DESCRIPTION
@@ -37,7 +37,6 @@ Imports:
 Suggests:
     bookdown,
     cyclocomp,
-    data.table,
     jsonlite,
     patrick (>= 0.2.0),
     rlang,
@@ -47,6 +46,8 @@ Suggests:
     tibble,
     tufte,
     withr (>= 2.5.0)
+Enhances:
+    data.table
 VignetteBuilder: 
     knitr
 Config/Needs/website: tidyverse/tidytemplate
