### 2024-12-04 f00aca6
Deploying to gh-pages from @ r-lib/lintr@c3d2935d355230e28ee79c0a541fd137d005bc22 🚀
