### 2024-12-08 #7240 @SaranjeetKaur
> Is {lintr} release the only current blocker, or are there other pending items?

@MichaelChirico No other pending items from me. Please feel free to prioritise as you see fit and drop a message to @lrnv here whenever you are ready to move to the publication process.
