### 2024-12-08 bfaf45a
Update NEWS.md
