### 2024-12-08 09f3fbd
Deploying to gh-pages from @ r-lib/lintr@edd299feccc78debc6c667407f7b122ea5e7c749 🚀
