### 2024-12-09 b835385
more tests, begin implementation
