### 2024-12-20 d8991c9
Include logo in the docs (#2700)
closes #1188
