### 2024-12-20 aafd047
Deploying to gh-pages from @ r-lib/lintr@d8991c9c22bc84cb44f44ca73a8dbe9a755ee929 🚀
