### 2025-01-15 53bdf7f
make loading the help DB a helper, and more robust
