+++ tests/testthat/helper.R
@@ -62,6 +62,16 @@ skip_if_not_utf8_locale <- function() {
   testthat::skip_if_not(l10n_info()[["UTF-8"]], "Not a UTF-8 locale")
 }
 
+safe_load_help_db <- function() {
+  help_db <- tryCatch(tools::Rd_db("lintr"), error = \(e) NULL)
+  # e.g. in dev under pkgload::load_all()
+  if (length(help_db) == 0L) {
+    help_db <- tryCatch(tools::Rd_db(dir = testthat::test_path("..", "..")), error = \(e) NULL)
+    testthat::skip_if_not(length(help_db) > 0L, message = "Package help corrupted or not installed")
+  }
+  help_db
+}
+
 pipes <- function(exclude = NULL) {
   if (getRversion() < "4.1.0") exclude <- unique(c(exclude, "|>"))
   all_pipes <- c(
