### 2025-01-15 6e8ea5f
adapt tests to upcoming {patrick} release
