+++ tests/testthat/helper.R
@@ -63,10 +63,10 @@ skip_if_not_utf8_locale <- function() {
 }
 
 safe_load_help_db <- function() {
-  help_db <- tryCatch(tools::Rd_db("lintr"), error = \(e) NULL)
+  help_db <- tryCatch(tools::Rd_db("lintr"), error = function(e) NULL)
   # e.g. in dev under pkgload::load_all()
   if (length(help_db) == 0L) {
-    help_db <- tryCatch(tools::Rd_db(dir = testthat::test_path("..", "..")), error = \(e) NULL)
+    help_db <- tryCatch(tools::Rd_db(dir = testthat::test_path("..", "..")), error = function(e) NULL)
     testthat::skip_if_not(length(help_db) > 0L, message = "Package help corrupted or not installed")
   }
   help_db
