### 2025-01-16 b6d945d
Adapt tests to upcoming {patrick} release (#2709)
