### 2025-01-16 cb4b885
Deploying to gh-pages from @ r-lib/lintr@b6d945d980e726029f89f96d7124b74541d8ee9a 🚀
