### 2025-01-28 c03a1be
initial implementation: just add the argument, deprecation, and add expected warnings
