+++ tests/testthat/test-assignment_linter.R
@@ -35,49 +35,60 @@ test_that("arguments handle <<- and ->/->> correctly", {
 
   # <<- is only blocked optionally
   expect_lint("1 <<- blah", NULL, linter)
-  expect_lint(
-    "1 <<- blah",
-    rex::rex("Replace <<- by assigning to a specific environment"),
-    assignment_linter(allow_cascading_assign = FALSE)
+  expect_warning(
+    expect_lint(
+      "1 <<- blah",
+      rex::rex("Replace <<- by assigning to a specific environment"),
+      assignment_linter(allow_cascading_assign = FALSE)
+    ),
+    "allow_cascading_assign"
   )
 
   # blocking -> can be disabled
-  expect_lint("1 -> blah", NULL, assignment_linter(allow_right_assign = TRUE))
-  expect_lint("1 ->> blah", NULL, assignment_linter(allow_right_assign = TRUE))
+  expect_warning(
+    expect_lint("1 -> blah", NULL, assignment_linter(allow_right_assign = TRUE)),
+    "allow_right_assign"
+  )
+  expect_warning(
+    expect_lint("1 ->> blah", NULL, assignment_linter(allow_right_assign = TRUE)),
+    "allow_right_assign"
+  )
   # blocked under cascading assign but not under right assign --> blocked
-  expect_lint(
-    "1 ->> blah",
-    lint_msg_right,
-    assignment_linter(allow_cascading_assign = FALSE, allow_right_assign = TRUE)
+  expect_warning(
+    expect_warning(
+      expect_lint(
+        "1 ->> blah",
+        lint_msg_right,
+        assignment_linter(allow_cascading_assign = FALSE, allow_right_assign = TRUE)
+      ),
+      "allow_cascading_assign"
+    ),
+    "allow_right_assign"
   )
 })
 
 test_that("arguments handle trailing assignment operators correctly", {
-  expect_lint("x <- y", NULL, assignment_linter(allow_trailing = FALSE))
-  expect_lint("foo(bar = 1)", NULL, assignment_linter(allow_trailing = FALSE))
+  linter <- assignment_linter(allow_trailing = FALSE)
+  expect_lint("x <- y", NULL, linter)
+  expect_lint("foo(bar = 1)", NULL, linter)
 
   expect_lint(
     "foo(bar =\n1)",
     rex::rex("= should not be trailing at the end of a line."),
-    assignment_linter(allow_trailing = FALSE)
+    linter
   )
 
-  expect_lint(
+  expect_lint("x <<-\ny", rex::rex("<<- should not be trailing"), linter)
+  expect_warning(
+    expect_lint(
     "x <<-\ny",
-    rex::rex("<<- should not be trailing"),
-    assignment_linter(allow_trailing = FALSE)
-  )
-  expect_lint(
-    "x <<-\ny",
-    rex::rex("Replace <<- by assigning to a specific environment"),
-    assignment_linter(allow_trailing = FALSE, allow_cascading_assign = FALSE)
+      rex::rex("Replace <<- by assigning to a specific environment"),
+      assignment_linter(allow_trailing = FALSE, allow_cascading_assign = FALSE)
+    ),
+    "allow_cascading_assign"
   )
 
-  expect_lint(
-    "x <- #Test \ny",
-    rex::rex("<- should not be trailing"),
-    assignment_linter(allow_trailing = FALSE)
-  )
+  expect_lint("x <- #Test \ny", rex::rex("<- should not be trailing"), linter)
 
   expect_lint(
     "is_long <-\nis %>%\ngather(measure, value, -Species) %>%\narrange(-value)",
@@ -87,7 +98,7 @@ test_that("arguments handle trailing assignment operators correctly", {
   expect_lint(
[…]
