### 2025-01-28 b29efda
replace allow* arguments with equivalent operator= logic
