+++ tests/testthat/test-assignment_linter.R
@@ -212,3 +212,121 @@ test_that("multiple lints throw correct messages", {
     "allow_cascading_assign"
   )
 })
+
+test_that("assignment operator can be toggled", {
+  eq_linter <- assignment_linter(operator = "=")
+  any_linter <- assignment_linter(operator = "any")
+  lint_message <- rex("Use =, not")
+
+  expect_lint("a = 1", NULL, eq_linter)
+  expect_lint("a = 1", NULL, any_linter)
+
+  expect_lint("a <- 1", lint_message, eq_linter)
+  expect_lint("a <- 1", NULL, any_linter)
+
+  expect_lint("a := 1", lint_message, eq_linter)
+  expect_lint("a := 1", NULL, any_linter)
+
+  expect_lint("a = 1; b <- 2", lint_message, eq_linter)
+  expect_lint("a = 1; b <- 2", NULL, any_linter)
+
+  expect_lint(
+    trim_some("
+      foo = function() {
+        a = 1
+      }
+    "),
+    NULL,
+    eq_linter
+  )
+  expect_lint(
+    trim_some("
+      foo = function() {
+        a = 1
+      }
+    "),
+    NULL,
+    any_linter
+  )
+
+  expect_lint(
+    trim_some("
+      foo = function() {
+        a <- 1
+      }
+    "),
+    list(lint_message, line_number = 3L),
+    eq_linter
+  )
+  expect_lint(
+    trim_some("
+      foo = function() {
+        a <- 1
+      }
+    "),
+    NULL,
+    any_linter
+  )
+
+  expect_lint("if ({a = TRUE}) 1", NULL, eq_linter)
+  expect_lint("if ({a = TRUE}) 1", NULL, any_linter)
+
+  expect_lint("if (a <- TRUE) 1", NULL, eq_linter)
+  expect_lint("if (a <- TRUE) 1", NULL, any_linter)
+
+  expect_lint("for (ii in {a = TRUE}) 1", NULL, eq_linter)
+  expect_lint("for (ii in {a = TRUE}) 1", NULL, any_linter)
+
+  expect_lint("for (ii in a <- TRUE) 1", NULL, eq_linter)
+  expect_lint("for (ii in a <- TRUE) 1", NULL, any_linter)
+
+  expect_lint("DT[, a := 1]", NULL, eq_linter)
+  expect_lint("DT[, a := 1]", NULL, any_linter)
+})
+
+test_that("multiple lints throw correct messages when both = and <- are allowed", {
+  expect_warning(
+    expect_lint(
+      trim_some("{
+        x <<- 1
+        y ->> 2
+        z -> 3
+        x %<>% as.character()
+        foo <- 1
+        bar = 2
+      }"),
+      list(
+        list(message = "Replace <<- by assigning to a specific environment", line_number = 2L),
+        list(message = "Replace ->> by assigning to a specific environment", line_number = 3L),
+        list(message = "Use <-, not ->", line_number = 4L),
+        list(message = "Avoid the assignment pipe %<>%", line_number = 5L)
+      ),
+      assignment_linter(allow_cascading_assign = FALSE, operator = "any")
+    ),
+    "allow_cascading_assign"
+  )
+})
+
+test_that("multiple lints throw correct messages when = is required", {
+  expect_warning(
+    expect_lint(
+      trim_some("{
+        x <<- 1
+        y ->> 2
+        z -> 3
+        x %<>% as.character()
+        foo <- 1
+        bar = 2
+      }"),
+      list(
+        list(message = "Replace <<- by assigning to a specific environment", line_number = 2L),
+        list(message = "Replace ->> by assigning to a specific environment", line_number = 3L),
+        list(message = "Use <-, not ->", line_number = 4L),
[…]
