+++ tests/testthat/test-assignment_linter.R
@@ -9,7 +9,7 @@ test_that("assignment_linter skips allowed usages", {
 
 test_that("assignment_linter blocks disallowed usages", {
   linter <- assignment_linter()
-  lint_msg <- rex::rex("Use <-, not =, for assignment.")
+  lint_msg <- rex::rex("Use one of <-, <<- for assignment, not =.")
 
   expect_lint("blah=1", lint_msg, linter)
   expect_lint("blah = 1", lint_msg, linter)
@@ -30,7 +30,7 @@ test_that("arguments handle <<- and ->/->> correctly", {
   linter <- assignment_linter()
   lint_msg_right <- rex::rex("Replace ->> by assigning to a specific environment")
 
-  expect_lint("1 -> blah", rex::rex("Use <-, not ->, for assignment."), linter)
+  expect_lint("1 -> blah", rex::rex("Use one of <-, <<- for assignment, not ->."), linter)
   expect_lint("1 ->> blah", lint_msg_right, linter)
 
   # <<- is only blocked optionally
@@ -82,7 +82,10 @@ test_that("arguments handle trailing assignment operators correctly", {
   expect_warning(
     expect_lint(
     "x <<-\ny",
-      rex::rex("Replace <<- by assigning to a specific environment"),
+      list(
+        rex::rex("Replace <<- by assigning to a specific environment"),
+        rex::rex("Assignment <<- should not be trailing")
+      ),
       assignment_linter(allow_trailing = FALSE, allow_cascading_assign = FALSE)
     ),
     "allow_cascading_assign"
@@ -103,12 +106,15 @@ test_that("arguments handle trailing assignment operators correctly", {
 
   expect_lint(
     "is %>%\ngather(measure, value, -Species) %>%\narrange(-value) ->\nis_long",
-    rex::rex("Use <-, not ->"),
+    rex::rex("Use one of <-, <<- for assignment, not ->"),
     assignment_linter()
   )
   expect_lint(
     "is %>%\ngather(measure, value, -Species) %>%\narrange(-value) ->\nis_long",
-    rex::rex("Use <-, not ->"),
+    list(
+      rex::rex("Use one of <-, <<- for assignment, not ->"),
+      rex::rex("Assignment -> should not be trailing")
+    ),
     linter
   )
   expect_warning(
@@ -189,7 +195,14 @@ test_that("%<>% throws a lint", {
   )
 
   # interaction with allow_trailing
-  expect_lint("x %<>%\n  sum()", "Assignment %<>% should not be trailing", assignment_linter(allow_trailing = FALSE))
+  expect_lint(
+    "x %<>%\n  sum()",
+    list(
+      "Avoid the assignment pipe %<>%",
+      "Assignment %<>% should not be trailing"
+    ),
+    assignment_linter(allow_trailing = FALSE)
+  )
 })
 
 test_that("multiple lints throw correct messages", {
@@ -204,7 +217,7 @@ test_that("multiple lints throw correct messages", {
       list(
         list(message = "Replace <<- by assigning to a specific environment", line_number = 2L),
         list(message = "Replace ->> by assigning to a specific environment", line_number = 3L),
-        list(message = "Use <-, not ->", line_number = 4L),
+        list(message = "Use <- for assignment, not ->", line_number = 4L),
         list(message = "Avoid the assignment pipe %<>%", line_number = 5L)
       ),
       assignment_linter(allow_cascading_assign = FALSE)
@@ -216,7 +229,7 @@ test_that("multiple lints throw correct messages", {
 test_that("assignment operator can be toggled", {
   eq_linter <- assignment_linter(operator = "=")
   any_linter <- assignment_linter(operator = "any")
-  lint_message <- rex("Use =, not")
+  lint_message <- rex("Use = for assignment, not")
 
[…]
