### 2025-01-28 c25bfde
simplify, remove := handling as before
