### 2025-01-28 4f39f6e
add docs, readability
