+++ man/assignment_linter.Rd
@@ -5,6 +5,7 @@
 \title{Assignment linter}
 \usage{
 assignment_linter(
+  operator = c("<-", "<<-"),
   allow_cascading_assign = TRUE,
   allow_right_assign = FALSE,
   allow_trailing = TRUE,
@@ -12,17 +13,20 @@ assignment_linter(
 )
 }
 \arguments{
-\item{allow_cascading_assign}{Logical, default \code{TRUE}.
+\item{operator}{Character vector of valid assignment operators. Defaults to allowing \verb{<-} and \verb{<<-}; other valid
+options are \code{=}, \verb{->}, \verb{->>}, and \verb{\%<>\%}.}
+
+\item{allow_cascading_assign}{(Deprecated) Logical, default \code{TRUE}.
 If \code{FALSE}, \code{\link[base:assignOps]{<<-}} and \verb{->>} are not allowed.}
 
-\item{allow_right_assign}{Logical, default \code{FALSE}. If \code{TRUE}, \verb{->} and \verb{->>} are allowed.}
+\item{allow_right_assign}{(Deprecated) Logical, default \code{FALSE}. If \code{TRUE}, \verb{->} and \verb{->>} are allowed.}
 
 \item{allow_trailing}{Logical, default \code{TRUE}. If \code{FALSE} then assignments aren't allowed at end of lines.}
 
-\item{allow_pipe_assign}{Logical, default \code{FALSE}. If \code{TRUE}, magrittr's \verb{\%<>\%} assignment is allowed.}
+\item{allow_pipe_assign}{(Deprecated) Logical, default \code{FALSE}. If \code{TRUE}, magrittr's \verb{\%<>\%} assignment is allowed.}
 }
 \description{
-Check that \verb{<-} is always used for assignment.
+Check that the specified operator is used for assignment.
 }
 \examples{
 # will produce lints
@@ -41,6 +45,11 @@ lint(
 lint(
   text = "x \%<>\% as.character()",
   linters = assignment_linter()
+)
+
+lint(
+  text = "x <- 1",
+  linters = assignment_linter(operator = "=")
 )
 
 # okay
@@ -78,6 +87,11 @@ lint(
 lint(
   text = "x \%<>\% as.character()",
   linters = assignment_linter(allow_pipe_assign = TRUE)
+)
+
+lint(
+  text = "x = 1",
+  linters = assignment_linter(operator = "=")
 )
 
 }
