### 2025-01-28 73e3cac
allow_trailing=FALSE lints when EQ_ASSIGN used
