+++ tests/testthat/test-expect_lint.R
@@ -3,7 +3,7 @@
 # for failure, always put the lint check or lint field that must fail first.
 
 linter <- assignment_linter()
-lint_msg <- "Use <-, not ="
+lint_msg <- "Use one of <-, <<- for assignment, not ="
 
 test_that("no checks", {
   expect_success(expect_no_lint("a", linter))
