### 2025-02-03 9e216c9
Duplicate tests to separate deprecated vs. new behavior
