+++ tests/testthat/test-assignment_linter.R
@@ -28,6 +28,7 @@ test_that("assignment_linter blocks disallowed usages", {
 
 test_that("arguments handle <<- and ->/->> correctly", {
   linter <- assignment_linter()
+  linter_yes_right <- assignment_linter(operator = c("->", "->>"))
   lint_msg_right <- rex::rex("Replace ->> by assigning to a specific environment")
 
   expect_lint("1 -> blah", rex::rex("Use one of <-, <<- for assignment, not ->."), linter)
@@ -35,35 +36,20 @@ test_that("arguments handle <<- and ->/->> correctly", {
 
   # <<- is only blocked optionally
   expect_lint("1 <<- blah", NULL, linter)
-  expect_warning(
-    expect_lint(
-      "1 <<- blah",
-      rex::rex("Replace <<- by assigning to a specific environment"),
-      assignment_linter(allow_cascading_assign = FALSE)
-    ),
-    "allow_cascading_assign"
+  expect_lint(
+    "1 <<- blah",
+    rex::rex("Replace <<- by assigning to a specific environment"),
+    assignment_linter(operator = "<-")
   )
 
   # blocking -> can be disabled
-  expect_warning(
-    expect_lint("1 -> blah", NULL, assignment_linter(allow_right_assign = TRUE)),
-    "allow_right_assign"
-  )
-  expect_warning(
-    expect_lint("1 ->> blah", NULL, assignment_linter(allow_right_assign = TRUE)),
-    "allow_right_assign"
-  )
-  # blocked under cascading assign but not under right assign --> blocked
-  expect_warning(
-    expect_warning(
-      expect_lint(
-        "1 ->> blah",
-        lint_msg_right,
-        assignment_linter(allow_cascading_assign = FALSE, allow_right_assign = TRUE)
-      ),
-      "allow_cascading_assign"
-    ),
-    "allow_right_assign"
+  expect_lint("1 -> blah", NULL, linter_yes_right)
+  expect_lint("1 ->> blah", NULL, linter_yes_right)
+  # we can also differentiate '->' and '->>'
+  expect_lint(
+    "1 ->> blah",
+    lint_msg_right,
+    assignment_linter(operator = c("<-", "->"))
   )
 })
 
@@ -79,16 +65,13 @@ test_that("arguments handle trailing assignment operators correctly", {
   )
 
   expect_lint("x <<-\ny", rex::rex("<<- should not be trailing"), linter)
-  expect_warning(
-    expect_lint(
-      "x <<-\ny",
-      list(
-        rex::rex("Replace <<- by assigning to a specific environment"),
-        rex::rex("Assignment <<- should not be trailing")
-      ),
-      assignment_linter(allow_trailing = FALSE, allow_cascading_assign = FALSE)
+  expect_lint(
+    "x <<-\ny",
+    list(
+      rex::rex("Replace <<- by assigning to a specific environment"),
+      rex::rex("Assignment <<- should not be trailing")
     ),
-    "allow_cascading_assign"
+    assignment_linter(operator = "<-", allow_trailing = FALSE)
   )
 
   expect_lint("x <- #Test \ny", rex::rex("<- should not be trailing"), linter)
@@ -117,13 +100,10 @@ test_that("arguments handle trailing assignment operators correctly", {
     ),
     linter
   )
-  expect_warning(
-    expect_lint(
-      "is %>%\ngather(measure, value, -Species) %>%\narrange(-value) ->\nis_long",
-      rex::rex("-> should not be trailing"),
-      assignment_linter(allow_right_assign = TRUE, allow_trailing = FALSE)
-    ),
-    "allow_right_assign"
+  expect_lint(
+    "is %>%\ngather(measure, value, -Species) %>%\narrange(-value) ->\nis_long",
+    rex::rex("-> should not be trailing"),
[…]
