### 2025-02-03 caf576c
Also warn for positionally-passed allow_cascadign_assign
