### 2025-02-03 662542a
More useful advice when cascading assignment is sometimes OK
