### 2025-02-03 a2af76b
Remove '<<-' from valid operators for this test
