### 2025-02-03 5202604
Use trim_some() everywhere
