### 2025-02-06 2aeede8
fully deprecate is.logical(operator)
