### 2025-02-06 fe3cd18
Initial progress migrating style for $xml_find_function_calls
