+++ R/condition_call_linter.R
@@ -77,7 +77,7 @@ condition_call_linter <- function(display_call = FALSE) {
     msg_fmt <- "Use %s(., call. = FALSE) not to display the call in an error message."
   }
 
-  xpath <- glue::glue("parent::expr[{call_cond}]/parent::expr")
+  xpath <- glue::glue("./self::*[{call_cond}]/parent::expr")
 
   Linter(linter_level = "expression", function(source_expression) {
     xml_calls <- source_expression$xml_find_function_calls(c("stop", "warning"))
