### 2025-02-06 3ca2af1
rest of first pass on linters, docs
