+++ R/consecutive_assertion_linter.R
@@ -35,24 +35,21 @@ consecutive_assertion_linter <- function() {
   next_expr <- "following-sibling::*[self::expr or self::expr_or_assign_or_help or self::equal_assign][1]"
 
   stopifnot_xpath <- glue("
-  parent::expr
-    /parent::expr[
-      expr[1]/SYMBOL_FUNCTION_CALL = {next_expr}/expr[1]/SYMBOL_FUNCTION_CALL
-    ]
-  ")
+  parent::expr[
+    expr[1]/SYMBOL_FUNCTION_CALL = {next_expr}/expr[1]/SYMBOL_FUNCTION_CALL
+  ]")
   assert_that_xpath <- glue("
-  parent::expr
-    /parent::expr[
-      not(SYMBOL_SUB[text() = 'msg'])
-      and not(following-sibling::expr[1]/SYMBOL_SUB[text() = 'msg'])
-      and expr[1]/SYMBOL_FUNCTION_CALL = {next_expr}/expr[1]/SYMBOL_FUNCTION_CALL
-    ]
-  ")
+  parent::expr[
+    not(SYMBOL_SUB[text() = 'msg'])
+    and not(following-sibling::expr[1]/SYMBOL_SUB[text() = 'msg'])
+    and expr[1]/SYMBOL_FUNCTION_CALL = {next_expr}/expr[1]/SYMBOL_FUNCTION_CALL
+  ]")
 
   Linter(linter_level = "file", function(source_expression) {
     # need the full file to also catch usages at the top level
     stopifnot_calls <- source_expression$xml_find_function_calls("stopifnot")
     assert_that_calls <- source_expression$xml_find_function_calls("assert_that")
+    # browser()
     bad_expr <- combine_nodesets(
       xml_find_all(stopifnot_calls, stopifnot_xpath),
       xml_find_all(assert_that_calls, assert_that_xpath)
