### 2025-02-07 8ca4ad3
Add a test of a lint for operator="any" case
