### 2025-02-09 fe958e4
assignment_linter() gains operator= argument to allow, e.g. '=' assignments (#2711)
* initial implementation: just add the argument, deprecation, and add expected warnings

* replace allow* arguments with equivalent operator= logic

* some progress unmangling

* fix most tests

* passing existing tests

* simplify, remove := handling as before

* add docs, readability

* NEWS, and mention "any" in man

* allow_trailing=FALSE lints when EQ_ASSIGN used

* fix tests

* last two tests

* delint

* Duplicate tests to separate deprecated vs. new behavior

* Also warn for positionally-passed allow_ca
[…]
