### 2025-02-11 39f7a59
Make $xml_find_function_calls() land on parent::expr instead (#2715)
* Initial progress migrating style for $xml_find_function_calls

* rest of first pass on linters, docs

* document updated to man/

* vestigial debugging comment

* ws change for consistency with other edits

---------

Co-authored-by: AshesITR <alexander.rosenstock@web.de>
