+++ R/length_levels_linter.R
@@ -22,7 +22,6 @@ length_levels_linter <- make_linter_from_function_xpath(
   function_names = "levels",
   xpath = "
   parent::expr
-    /parent::expr
     /parent::expr[expr/SYMBOL_FUNCTION_CALL[text() = 'length']]
   ",
   lint_message = "nlevels(x) is better than length(levels(x))."
