+++ .devcontainer/r-devel/Dockerfile
@@ -1,0 +1,14 @@
+FROM docker.io/rhub/r-minimal:devel
+
+RUN apk update && \
+  apk add --no-cache git gcc g++ libxml2 linux-headers musl-dev
+
+COPY DESCRIPTION .
+
+RUN Rscript -e '                           \
+  install.packages("remotes");             \
+  remotes::install_deps(dependencies = c(  \
+    "Imports",                             \
+    "Config/needs/development"             \
+  ))                                       \
+'
