### 2025-02-11 58cf978
Switch to r-devel, minimal image not suitable
