### 2025-02-11 40f5c67
rocker/r-devel extremely heavy & slow to load
