+++ .devcontainer/r-devel/Dockerfile
@@ -1,4 +1,4 @@
-FROM FROM registry.gitlab.com/jangorecki/dockerfiles/r-devel-gcc
+FROM registry.gitlab.com/jangorecki/dockerfiles/r-devel-gcc
 
 RUN apt-get -qq update && \
   apt-get install -y --no-install-recommends git libxml2-dev
