### 2025-02-11 97870a0
Revert "to-be-deleted: massive GHA for all versions"
This reverts commit b7a34d173f35f784b826ef9f229e2638d44d7fed.
