+++ .github/workflows/R-CMD-check.yaml
@@ -26,72 +26,20 @@ jobs:
       fail-fast: false
       matrix:
         config:
-          - {os: windows-latest,  r: '4.0.0'}
-          - {os: windows-latest,  r: '4.0.1'}
-          - {os: windows-latest,  r: '4.0.2'}
-          - {os: windows-latest,  r: '4.0.3'}
-          - {os: windows-latest,  r: '4.0.4'}
-          - {os: windows-latest,  r: '4.0.5'}
-          - {os: windows-latest,  r: '4.1.0'}
-          - {os: windows-latest,  r: '4.1.1'}
-          - {os: windows-latest,  r: '4.1.2'}
-          - {os: windows-latest,  r: '4.1.3'}
-          - {os: windows-latest,  r: '4.2.0'}
-          - {os: windows-latest,  r: '4.2.1'}
-          - {os: windows-latest,  r: '4.2.2'}
-          - {os: windows-latest,  r: '4.2.3'}
-          - {os: windows-latest,  r: '4.3.0'}
-          - {os: windows-latest,  r: '4.3.1'}
-          - {os: windows-latest,  r: '4.3.2'}
-          - {os: windows-latest,  r: '4.3.3'}
-          - {os: windows-latest,  r: '4.4.0'}
-          - {os: windows-latest,  r: '4.4.1'}
-          - {os: windows-latest,  r: '4.4.2'}
-          - {os: windows-latest,  r: 'devel'}
-          - {os: macos-latest,  r: '4.0.0'}
-          - {os: macos-latest,  r: '4.0.1'}
-          - {os: macos-latest,  r: '4.0.2'}
-          - {os: macos-latest,  r: '4.0.3'}
-          - {os: macos-latest,  r: '4.0.4'}
-          - {os: macos-latest,  r: '4.0.5'}
-          - {os: macos-latest,  r: '4.1.0'}
-          - {os: macos-latest,  r: '4.1.1'}
-          - {os: macos-latest,  r: '4.1.2'}
-          - {os: macos-latest,  r: '4.1.3'}
-          - {os: macos-latest,  r: '4.2.0'}
-          - {os: macos-latest,  r: '4.2.1'}
-          - {os: macos-latest,  r: '4.2.2'}
-          - {os: macos-latest,  r: '4.2.3'}
-          - {os: macos-latest,  r: '4.3.0'}
-          - {os: macos-latest,  r: '4.3.1'}
-          - {os: macos-latest,  r: '4.3.2'}
-          - {os: macos-latest,  r: '4.3.3'}
-          - {os: macos-latest,  r: '4.4.0'}
-          - {os: macos-latest,  r: '4.4.1'}
-          - {os: macos-latest,  r: '4.4.2'}
-          - {os: macos-latest,  r: 'devel'}
-          - {os: ubuntu-latest,  r: '4.0.0'}
-          - {os: ubuntu-latest,  r: '4.0.1'}
-          - {os: ubuntu-latest,  r: '4.0.2'}
-          - {os: ubuntu-latest,  r: '4.0.3'}
-          - {os: ubuntu-latest,  r: '4.0.4'}
-          - {os: ubuntu-latest,  r: '4.0.5'}
-          - {os: ubuntu-latest,  r: '4.1.0'}
-          - {os: ubuntu-latest,  r: '4.1.1'}
-          - {os: ubuntu-latest,  r: '4.1.2'}
-          - {os: ubuntu-latest,  r: '4.1.3'}
-          - {os: ubuntu-latest,  r: '4.2.0'}
-          - {os: ubuntu-latest,  r: '4.2.1'}
-          - {os: ubuntu-latest,  r: '4.2.2'}
[…]
