### 2025-02-12 #7240 @IndrajeetPatil
@lrnv The new version of `{lintr}` is now [on CRAN](https://cran.r-project.org/web/packages/lintr/index.html), so there is no longer any blocker from our end for accepting this submission 😄
