### 2025-02-12 b96e5e5
3.2.0 as submitted to CRAN (#2726)
* Create a second r-devel devcontainer, set it as default

* Need repos set on this image

* DESCRIPTION version

* Copy-edit to NEWS

* empty entry for backports:4.4.0

* Fix up backport_linter for oldrel stuff

* ignore references to removed linter

* to-be-deleted: massive GHA for all versions

* Revert "to-be-deleted: massive GHA for all versions"

This reverts commit b7a34d173f35f784b826ef9f229e2638d44d7fed.

* Fix up exclusions of large files
