### 2025-02-12 4774c46
Deprecate passing non-linter functions, names
