### 2025-02-12 63f5b08
Deprecate linter= argument of Lint()
