+++ tests/testthat/test-lint.R
@@ -156,14 +156,15 @@ test_that("exclusions work with custom linter names", {
 })
 
 test_that("old compatibility usage errors", {
+  err = rex::rex("Expected `", anything, "` to be a function of class <linter>")
+
   expect_error(
     expect_lint(
       "a == NA",
       "Use is.na",
       linters = equals_na_linter
     ),
-    regexp = "Passing linters as variables",
-    fixed = TRUE
+    err
   )
 
   expect_error(
@@ -172,8 +173,7 @@ test_that("old compatibility usage errors", {
       "Use <-",
       linters = assignment_linter
     ),
-    regexp = "Passing linters as variables",
-    fixed = TRUE
+    err
   )
 
   # Also within `linters_with_defaults()` (#1725)
@@ -183,8 +183,7 @@ test_that("old compatibility usage errors", {
       "Use <-",
       linters = linters_with_defaults(assignment_linter)
     ),
-    regexp = "Passing linters as variables",
-    fixed = TRUE
+    err
   )
 
   expect_error(
@@ -193,8 +192,7 @@ test_that("old compatibility usage errors", {
       "Use is.na",
       linters = unclass(equals_na_linter())
     ),
-    regexp = "The use of linters of class 'function'",
-    fixed = TRUE
+    err
   )
 
   # Trigger compatibility in auto_names()
@@ -204,20 +202,17 @@ test_that("old compatibility usage errors", {
       "Use is.na",
       linters = list(unclass(equals_na_linter()))
     ),
-    "The use of linters of class 'function'",
-    fixed = TRUE
+    err
   )
 
   expect_error(
     lint("a <- 1\n", linters = function(two, arguments) NULL),
-    regexp = "The use of linters of class 'function'",
-    fixed = TRUE
+    err
   )
 
   expect_error(
     lint("a <- 1\n", linters = "equals_na_linter"),
-    regexp = "Expected `linters()` to be a function of class <linter>",
-    fixed = TRUE
+    err
   )
 })
 
