### 2025-02-12 d5eada5
Remove with_defaults
