+++ man/lintr-deprecated.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/lintr-deprecated.R, R/with.R
+% Please edit documentation in R/lintr-deprecated.R
 \name{lintr-deprecated}
 \alias{lintr-deprecated}
 \alias{closed_curly_linter}
@@ -12,7 +12,6 @@
 \alias{no_tab_linter}
 \alias{extraction_operator_linter}
 \alias{unnecessary_nested_if_linter}
-\alias{with_defaults}
 \title{Deprecated functions in lintr}
 \usage{
 closed_curly_linter(allow_single_line = FALSE)
@@ -34,8 +33,6 @@ no_tab_linter()
 extraction_operator_linter()
 
 unnecessary_nested_if_linter()
-
-with_defaults(..., default = default_linters)
 }
 \arguments{
 \item{allow_single_line, semicolon}{Irrelevant parameters to defunct linters.}
