+++ NAMESPACE
@@ -163,7 +163,6 @@ export(use_lintr)
 export(vector_logic_linter)
 export(which_grepl_linter)
 export(whitespace_linter)
-export(with_defaults)
 export(with_id)
 export(xml_nodes_to_lints)
 export(xp_call_name)
