### 2025-02-12 32499c7
progress deprecation of extraction_operator and unnecessary_nested_if linters
