### 2025-02-12 31e3d75
Progress deprecation of assignment_linter() args
