### 2025-02-12 644e696
Deprecate linter= argument of Lint() (#2730)
