### 2025-02-12 13563a1
clearer variable
