### 2025-02-12 e3db2e5
Remove check_dots() checks of old usage patterns
