### 2025-02-12 ae93465
Remove 4 defunct linters: {closed,open}_curly_linter(), paren_brace_linter(), semicolon_terminator_linter() (#2732)
* Remove 4 defunct linters

* remove "defunct" entry, too
