+++ NAMESPACE
@@ -29,7 +29,6 @@ export(brace_linter)
 export(checkstyle_output)
 export(class_equals_linter)
 export(clear_cache)
-export(closed_curly_linter)
 export(commas_linter)
 export(commented_code_linter)
 export(comparison_negation_linter)
@@ -111,11 +110,9 @@ export(object_name_linter)
 export(object_overwrite_linter)
 export(object_usage_linter)
 export(one_call_pipe_linter)
-export(open_curly_linter)
 export(outer_negation_linter)
 export(package_hooks_linter)
 export(paren_body_linter)
-export(paren_brace_linter)
 export(paste_linter)
 export(pipe_call_linter)
 export(pipe_consistency_linter)
@@ -134,7 +131,6 @@ export(sample_int_linter)
 export(sarif_output)
 export(scalar_in_linter)
 export(semicolon_linter)
-export(semicolon_terminator_linter)
 export(seq_linter)
 export(single_quotes_linter)
 export(sort_linter)
