### 2025-02-12 580156d
Remove error about incorrect ways to pass linters (#2729)
* Deprecate passing non-linter functions, names

* update tests

* delint

* clearer variable
