### 2025-02-13 50fa35c
Remove with_defaults() (#2731)
* Remove with_defaults

* roxygenize

* fix tests
