+++ man/lintr-deprecated.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/lintr-deprecated.R, R/with.R
+% Please edit documentation in R/lintr-deprecated.R
 \name{lintr-deprecated}
 \alias{lintr-deprecated}
 \alias{unneeded_concatenation_linter}
@@ -8,7 +8,6 @@
 \alias{no_tab_linter}
 \alias{extraction_operator_linter}
 \alias{unnecessary_nested_if_linter}
-\alias{with_defaults}
 \title{Deprecated functions in lintr}
 \usage{
 unneeded_concatenation_linter(allow_single_expression = TRUE)
@@ -22,8 +21,6 @@ no_tab_linter()
 extraction_operator_linter()
 
 unnecessary_nested_if_linter()
-
-with_defaults(..., default = default_linters)
 }
 \description{
 Deprecated functions in lintr
