### 2025-02-13 #7240 @IndrajeetPatil
@MichaelChirico I tried creating an archive release on Zenodo, but I don't see r-lib organization repos in my Github-synced Zenodo account (https://zenodo.org/account/settings/github/).

IINM, Zenodo needs to be approved for "Third-party application access policy" for this to happen, but I don't have the privileges to change organization settings: https://github.com/organizations/r-lib/settings/oauth_application_policy

The alternative is to create such a release manually, but that's too much of a manual and error-prone labour.
