### 2025-02-13 f725e36
Progress deprecation of extraction_operator and unnecessary_nested_if linters (#2733)
* progress deprecation of extraction_operator and unnecessary_nested_if linters

* reset tags too

* roxygnize updated metadata
