+++ man/brace_linter.Rd
@@ -12,9 +12,9 @@ brace_linter(
 \arguments{
 \item{allow_single_line}{If \code{TRUE}, allow an open and closed curly pair on the same line.}
 
-\item{function_bodies}{Whether to require function bodies to be wrapped in curly braces. One of
+\item{function_bodies}{When to require function bodies to be wrapped in curly braces. One of
 \itemize{
-\item \code{"always"} to require braces for all function definitions, including inline functions,
+\item \code{"always"} to require braces around all function bodies, including inline functions,
 \item \code{"not_inline"} to require braces when a function body does not start on the same line as its signature,
 \item \code{"multi_line"} (the default) to require braces when a function definition spans multiple lines,
 \item \code{"never"} to never require braces in function bodies.
