### 2025-02-13 c415fee
docs: fix NEWS
