### 2025-02-13 e21c4e7
Revert "use 'true' over '1' for boolean literal in XPath"
This reverts commit 4223eea4d22b3704a5a4e04d3230f3430113bc35.
