+++ tests/testthat/test-brace_linter.R
@@ -313,10 +313,10 @@ test_that("brace_linter lints function expressions correctly", {
       x + 4
     }
   ")
-  expect_lint(lines, NULL, linter_always)
-  expect_lint(lines, NULL, linter_multi_line)
-  expect_lint(lines, NULL, linter_not_inline)
-  expect_lint(lines, NULL, linter_never)
+  expect_no_lint(lines, linter_always)
+  expect_no_lint(lines, linter_multi_line)
+  expect_no_lint(lines, linter_not_inline)
+  expect_no_lint(lines, linter_never)
 
   lints_single_line <- list(
     rex::rex("Opening curly braces should never go on their own line and should always be followed by a new line."),
@@ -326,10 +326,9 @@ test_that("brace_linter lints function expressions correctly", {
   expect_lint("function(x) { x + 4 }", lints_single_line, linter_multi_line)
   expect_lint("function(x) { x + 4 }", lints_single_line, linter_not_inline)
   expect_lint("function(x) { x + 4 }", lints_single_line, linter_never)
-  # using function_bodies = "always" without allow_single_line = TRUE prohibits inline function definitions:
-  expect_lint(
+  # using function_bodies = "always" should only prohibit inline function definitions with allow_single_line = FALSE (the default):
+  expect_no_lint(
     "function(x) { x + 4 }",
-    NULL,
     brace_linter(allow_single_line = TRUE, function_bodies = "always")
   )
 
