### 2025-02-13 689f803
test: for missing newline after opening brace
or closing brace not on sep line
