### 2025-02-13 153adef
test: for fn body wrapped in additional unneeded parentheses
