### 2025-02-13 543512f
Allow that a tag has no associated linters (esp. "deprecated")
