### 2025-02-13 31c94c5
avoid using linter_auto_name, add tests
