+++ R/lintr-package.R
@@ -21,5 +21,6 @@ NULL
 
 # make binding available for mock testing
 # ref: https://testthat.r-lib.org/dev/reference/local_mocked_bindings.html#base-functions
+requireNamespace <- NULL
 unlink <- NULL
 quit <- NULL
