### 2025-02-13 3711615
move NEWS entry
