+++ R/lintr-package.R
@@ -21,6 +21,8 @@ NULL
 
 # make binding available for mock testing
 # ref: https://testthat.r-lib.org/dev/reference/local_mocked_bindings.html#base-functions
+# nolint start: object_name_linter. These will be copied from base style.
 requireNamespace <- NULL
 unlink <- NULL
 quit <- NULL
+# nolint end: object_name_linter.
