### 2025-02-13 5982a8b
Make require_testthat helper function (#2587)
* Make require_testthat helper function

* Update NEWS.md

* Add back testthat scopeing

* Roxygenise

* Add space in stop message

* Add a default to require_testthat

* avoid using linter_auto_name, add tests

* move NEWS entry

* I am in the wrong project. Notes is right indeed :)

* delint

* nolint false positive

---------

Co-authored-by: Michael Chirico <chiricom@google.com>
Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>
