### 2025-02-13 103840f
Slight refactor
