### 2025-02-13 5e427c9
Add param `function_bodies` to `brace_linter()` (#2240)
* Add param `function_braces` to `brace_linter()`

* refactor, add option not_inline

* update NEWS bullet

* Fix typos

* Roxygenize

* Move NEWS item, clarify it

* Update R/brace_linter.R

Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>

* Update R/brace_linter.R

Co-authored-by: Michael Chirico <michaelchirico4@gmail.com>

* Use list in roxygen doc

* Use 'true' in XPath for clarity

* rename param `function_braces` to `function_bodies`

* Revert "Use 'true' in XPath for clarity"

This reverts commit 5eaae61e1e3f023d3db4e0
[…]
