### 2025-02-13 e18302d
Remove vestiges of R<4 support
