+++ tests/testthat/test-condition_message_linter.R
@@ -1,18 +1,18 @@
 test_that("condition_message_linter skips allowed usages", {
   linter <- condition_message_linter()
 
-  expect_lint("stop('a string', 'another')", NULL, linter)
-  expect_lint("warning('a string', 'another')", NULL, linter)
-  expect_lint("message('a string', 'another')", NULL, linter)
+  expect_no_lint("stop('a string', 'another')", linter)
+  expect_no_lint("warning('a string', 'another')", linter)
+  expect_no_lint("message('a string', 'another')", linter)
   # extracted calls likely don't obey base::stop() semantics
-  expect_lint("ctx$stop(paste('a', 'b'))", NULL, linter)
-  expect_lint("ctx@stop(paste('a', 'b'))", NULL, linter)
+  expect_no_lint("ctx$stop(paste('a', 'b'))", linter)
+  expect_no_lint("ctx@stop(paste('a', 'b'))", linter)
 
   # sprintf is OK -- gettextf() enforcement is left to other linters
-  expect_lint("stop(sprintf('A %s!', 'string'))", NULL, linter)
+  expect_no_lint("stop(sprintf('A %s!', 'string'))", linter)
 
   # get multiple sep= in one expression
-  expect_lint(
+  expect_no_lint(
     trim_some("
       tryCatch(
         foo(x),
@@ -20,7 +20,6 @@ test_that("condition_message_linter skips allowed usages", {
         warning = function(w) warning(paste(a, b, sep = '--')),
       )
     "),
-    NULL,
     linter
   )
 })
@@ -28,9 +27,8 @@ test_that("condition_message_linter skips allowed usages", {
 skip_if_not_installed("tibble")
 patrick::with_parameters_test_that(
   "paste/paste0 allowed by condition_message_linter when using other seps and/or collapse",
-  expect_lint(
+  expect_no_lint(
     sprintf("%s(%s(x, %s = '%s'))", condition, fun, parameter, arg),
-    NULL,
     condition_message_linter()
   ),
   .cases = tibble::tribble(
@@ -48,36 +46,27 @@ patrick::with_parameters_test_that(
 )
 
 test_that("condition_message_linter blocks simple disallowed usages", {
-  expect_lint(
-    "stop(paste('a string', 'another'))",
-    rex::rex("Don't use paste to build stop strings."),
-    condition_message_linter()
-  )
+  linter <- condition_message_linter()
+  lint_msg_paste_stop <- rex::rex("Don't use paste to build stop strings.")
+
+  expect_lint("stop(paste('a string', 'another'))", lint_msg_paste_stop, linter)
 
   expect_lint(
     "warning(paste0('a string ', 'another'))",
     rex::rex("Don't use paste0 to build warning strings."),
-    condition_message_linter()
+    linter
   )
 
   # `sep` argument allowed, but only if it is different from default
-  expect_lint(
-    "stop(paste(x, sep = ' '))",
-    rex::rex("Don't use paste to build stop strings."),
-    condition_message_linter()
-  )
+  expect_lint("stop(paste(x, sep = ' '))", lint_msg_paste_stop, linter)
 
   # not thrown off by named arguments
-  expect_lint(
-    "stop(paste('a', 'b'), call. = FALSE)",
-    rex::rex("Don't use paste to build stop strings."),
-    condition_message_linter()
-  )
+  expect_lint("stop(paste('a', 'b'), call. = FALSE)", lint_msg_paste_stop, linter)
 
   expect_lint(
     "warning(paste0('a', 'b'), immediate. = TRUE)",
     rex::rex("Don't use paste0 to build warning strings."),
-    condition_message_linter()
+    linter
   )
 
   expect_lint(
[…]
