+++ R/implicit_integer_linter.R
@@ -47,9 +47,9 @@
 #' @export
 implicit_integer_linter <- function(allow_colon = FALSE) {
   if (allow_colon) {
-    xpath <- "//NUM_CONST[not(parent::expr[
-      parent::expr/OP-COLON
-      or parent::expr[OP-MINUS]/parent::expr/OP-COLON
+    xpath <- "//NUM_CONST[not(parent::expr/parent::expr[
+      OP-COLON
+      or (OP-MINUS and parent::expr/OP-COLON)
     ])]"
   } else {
     xpath <- "//NUM_CONST"
