### 2025-02-14 c2dce5c
Remove vestiges of R<4 support (#2750)
