### 2025-02-14 ba91608
Exempt OP-LAMBDA the same as FUNCTION in unnecessary_nesting_linter (#2749)
* Exempt OP-LAMBDA the same as FUNCTION

* NEWS, test
