### 2025-02-14 fd21b4d
Deploying to gh-pages from @ r-lib/lintr@ba916088d4b3aa078d91ac8d55b7043e7a8132e0 🚀
