### 2025-02-14 c2fb844
rm vestigial variables
