+++ tests/testthat/test-unnecessary_nesting_linter.R
@@ -470,14 +470,14 @@ test_that("unnecessary_nesting_linter skips allowed usages", {
     linter
   )
 
-  # before _or_ after the inner if, #2245
+  # '=' operator also (which uses non-<expr> node), #2245
   expect_no_lint(
     trim_some("
       if (x && a) {
+        y = x + 1L
         if (y || b) {
           1L
         }
-        y <- x + 1L
       }
     "),
     linter
