### 2025-02-14 b91f364
fixed: =3 for OP-RIGHT-BRACE
