### 2025-02-14 30441f4
Lint() always has message class character, not an object
