### 2025-02-14 904df7f
capture emitted output in tests
