### 2025-02-14 0cec2c1
also confirm invisible(), why not
