+++ tests/testthat/test-methods.R
@@ -134,7 +134,7 @@ test_that("print.lints works", {
 
   tmp <- withr::local_tempfile()
   stopifnot(file.create(tmp))
-  expect_message(print(lint(tmp)), "No lints found")
+  expect_message(expect_invisible(print(lint(tmp))), "No lints found")
 })
 
 test_that("split.lint works as intended", {
