### 2025-02-14 2e2cfbf
changing approach for linting 'message' in particular
