### 2025-02-14 503b0d4
Split undesirable_function_linter into always. vs call-only cases
