+++ R/utils.R
@@ -130,7 +130,7 @@ try_silently <- function(expr) {
 # interface to work like options() or setwd() -- returns the old value for convenience
 set_lang <- function(new_lang) {
   old_lang <- Sys.getenv("LANGUAGE", unset = NA)
-  Sys.setenv(LANGUAGE = new_lang) # nolint: undesirable_function. Avoiding {withr} dep in pkg.
+  Sys.setenv(LANGUAGE = new_lang) # nolint: undesirable_function_call. Avoiding {withr} dep in pkg.
   old_lang
 }
 # handle the logic of either unsetting if it was previously unset, or resetting
@@ -138,7 +138,7 @@ reset_lang <- function(old_lang) {
   if (is.na(old_lang)) {
     Sys.unsetenv("LANGUAGE")
   } else {
-    Sys.setenv(LANGUAGE = old_lang) # nolint: undesirable_function. Avoiding {withr} dep in pkg.
+    Sys.setenv(LANGUAGE = old_lang) # nolint: undesirable_function_call. Avoiding {withr} dep in pkg.
   }
 }
 
