### 2025-02-14 b60b993
Push validation into a helper
