+++ tests/testthat/test-string_boundary_linter.R
@@ -8,6 +8,10 @@ test_that("string_boundary_linter skips allowed grepl() usages", {
   # regex pattern --> no lint
   expect_no_lint("grepl('^[a-z]', x)", linter)
   expect_no_lint("grepl('[a-z]$', x)", linter)
+  # anchor-ish but not a regex --> no lint
+  expect_no_lint(R"[grepl("\\^", x)]", linter)
+  expect_no_lint(R"[grepl("\\\\^", x)]", linter)
+  expect_no_lint(R"[grepl("$\\\\", x)]", linter)
 
   # ignore.case --> no lint
   expect_no_lint("grepl('^abc', x, ignore.case = TRUE)", linter)
