### 2025-02-15 2ee4455
correct behavior in multi-pattern case too
