### 2025-02-15 f88df06
implicit_integer_linter(allow_colon=TRUE) ignores negative inputs (#2751)
