### 2025-02-17 15b75d8
Omit anchor-esque pattern in string_boundary_linter() (#2757)
