+++ tests/testthat/test-string_boundary_linter.R
@@ -8,6 +8,10 @@ test_that("string_boundary_linter skips allowed grepl() usages", {
   # regex pattern --> no lint
   expect_no_lint("grepl('^[a-z]', x)", linter)
   expect_no_lint("grepl('[a-z]$', x)", linter)
+  # anchor-ish but not a regex --> no lint (#2636)
+  expect_no_lint(R"[grepl("\\^", x)]", linter)
+  expect_no_lint(R"[grepl("\\\\^", x)]", linter)
+  expect_no_lint(R"[grepl("$\\\\", x)]", linter)
 
   # ignore.case --> no lint
   expect_no_lint("grepl('^abc', x, ignore.case = TRUE)", linter)
@@ -151,6 +155,8 @@ test_that("whole-string regex recommends ==, not {starts,ends}With()", {
 })
 
 test_that("vectorization + metadata work as intended", {
+  linter <- string_boundary_linter()
+
   expect_lint(
     trim_some("{
       substring(a, 1, 3) == 'abc'
@@ -176,6 +182,16 @@ test_that("vectorization + metadata work as intended", {
       list("endsWith", line_number = 10L),
       list("==", line_number = 11L)
     ),
-    string_boundary_linter()
+    linter
+  )
+
+  # some but not all anchor-esque as in #2636
+  expect_lint(
+    trim_some(R"[{
+      grepl("\\^", x)
+      grepl("^abc", x)
+    }]"),
+    list("startsWith", line_number = 3L),
+    linter
   )
 })
