### 2025-02-17 a370a9b
Ignore comments in empty call like foo() in missing_argument_linter (#2754)
* expect_no_lint

* New tests

* probable fix

* NEWS

---------

Co-authored-by: AshesITR <alexander.rosenstock@web.de>
