### 2025-02-17 1062623
Lint() always has message class character, not an object (#2755)
* Lint() always has message class character, not an object

* one more case

* changing approach for linting 'message' in particular

* Split undesirable_function_linter into always. vs call-only cases

* Push validation into a helper

* no trailing ','

* fix linting

---------

Co-authored-by: AshesITR <alexander.rosenstock@web.de>
