### 2025-02-17 18bd041
Skip when 2nd expression under if() is an '=' assignment in unnecessary_nesting_linter() (#2753)
* expect_no_lint

* new test case

* Untested fix

* fixed: =3 for OP-RIGHT-BRACE

* NEWS
