+++ tests/testthat/test-unnecessary_nesting_linter.R
@@ -2,7 +2,7 @@ test_that("unnecessary_nesting_linter skips allowed usages", {
   linter <- unnecessary_nesting_linter()
 
   # parallel stops() and return()s are OK
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (A) {
         stop()
@@ -10,11 +10,10 @@ test_that("unnecessary_nesting_linter skips allowed usages", {
         stop()
       }
     "),
-    NULL,
     linter
   )
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (A) {
         return()
@@ -22,14 +21,13 @@ test_that("unnecessary_nesting_linter skips allowed usages", {
         return()
       }
     "),
-    NULL,
     linter
   )
 })
 
 test_that("Multiple if/else statements don't require unnesting", {
   # with further branches, reducing nesting might be less readable
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (x == 'a') {
         stop()
@@ -39,26 +37,24 @@ test_that("Multiple if/else statements don't require unnesting", {
         stop()
       }
     "),
-    NULL,
     unnecessary_nesting_linter()
   )
 })
 
 test_that("else-less if statements don't lint", {
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (x == 4) {
         msg <- 'failed'
         stop(msg)
       }
     "),
-    NULL,
     unnecessary_nesting_linter()
   )
 })
 
 test_that("non-terminal expressions are not considered for the logic", {
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (x == 4) {
         x <- 5
@@ -67,13 +63,12 @@ test_that("non-terminal expressions are not considered for the logic", {
         return(x)
       }
     "),
-    NULL,
     unnecessary_nesting_linter()
   )
 })
 
 test_that("parallels in further nesting are skipped", {
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (length(bucket) > 1) {
         return(age)
@@ -86,7 +81,6 @@ test_that("parallels in further nesting are skipped", {
         }
       }
     "),
-    NULL,
     unnecessary_nesting_linter()
   )
 })
@@ -154,42 +148,39 @@ test_that("unnecessary_nesting_linter blocks if/else with one exit branch", {
   expect_lint(stop_warning_lines, lint_msg, linter)
 
   # Optionally consider 'warning' as an exit call --> no lint
-  expect_lint(stop_warning_lines, NULL, unnecessary_nesting_linter(branch_exit_calls = "warning"))
+  expect_no_lint(stop_warning_lines, unnecessary_nesting_linter(branch_exit_calls = "warning"))
 })
 
 test_that("unnecessary_nesting_linter skips one-line functions", {
   linter <- unnecessary_nesting_linter()
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       foo <- function(x) {
         return(x)
       }
     "),
-    NULL,
     linter
   )
 
   # purrr anonymous functions also get skipped
-  expect_lint(
+  expect_no_lint(
     trim_some("
       purrr::map(x, ~ {
         .x
       })
     "),
-    NULL,
     linter
   )
 
   # ditto short-hand lambda
   skip_if_not_r_version("4.1.0")
-  expect_lint(
+  expect_no_lint(
     trim_some("
       foo <- \\(x) {
         return(x)
       }
     "),
-    NULL,
     linter
   )
 })
@@ -197,30 +188,28 @@ test_that("unnecessary_nesting_linter skips one-line functions", {
[…]
