### 2025-02-17 330dacf
Allow that a tag has no associated linters (esp. "deprecated") (#2745)
* missed {} in cli evaluation

* Allow that a tag has no associated linters (esp. "deprecated")

* progress deprecation of 4 more linters (#2746)

* error, not warning

* Still need to be inside \section{}

* avoid 'body' overwrite with more descriptive name

* actually roxygenize; shrink diff

* dont bother with tags=NULL complexity

* typo

Co-authored-by: AshesITR <alexander.rosenstock@web.de>

---------

Co-authored-by: AshesITR <alexander.rosenstock@web.de>
