### 2025-02-18 #7240 @lrnv
@hadley You need to archive the last version of lintr on Zenodo. For that, you need to connect with the r-lib account on zenodo IMHO, and once you are there there is an automatic archival tool
