+++ R/seq_linter.R
@@ -66,6 +66,10 @@ seq_linter <- function() {
     ]
   ")
 
+  seq_len_xpath <- "
+    self::*[following-sibling::expr/expr/SYMBOL_FUNCTION_CALL[text() = 'length']]
+  "
+
   ## The actual order of the nodes is document order
   ## In practice we need to handle length(x):1
   get_fun <- function(expr, n) {
@@ -87,17 +91,17 @@ seq_linter <- function() {
     xml <- source_expression$xml_parsed_content
     seq_calls <- source_expression$xml_find_function_calls("seq")
 
-    badx <- combine_nodesets(
+    seq_expr <- combine_nodesets(
       xml_find_all(seq_calls, seq_xpath),
       xml_find_all(xml, colon_xpath)
     )
 
-    dot_expr1 <- get_fun(badx, 1L)
-    dot_expr2 <- get_fun(badx, 2L)
+    dot_expr1 <- get_fun(seq_expr, 1L)
+    dot_expr2 <- get_fun(seq_expr, 2L)
     seq_along_idx <- grepl("length(", dot_expr1, fixed = TRUE) | grepl("length(", dot_expr2, fixed = TRUE)
     rev_idx <- startsWith(dot_expr2, "1")
 
-    replacement <- rep("seq_along(...)", length(badx))
+    replacement <- rep("seq_along(...)", length(seq_expr))
     replacement[!seq_along_idx] <- paste0("seq_len(", ifelse(rev_idx, dot_expr1, dot_expr2)[!seq_along_idx], ")")
     replacement[rev_idx] <- paste0("rev(", replacement[rev_idx], ")")
 
@@ -113,6 +117,17 @@ seq_linter <- function() {
       )
     )
 
-    xml_nodes_to_lints(badx, source_expression, lint_message, type = "warning")
+    seq_lints <- xml_nodes_to_lints(seq_expr, source_expression, lint_message, type = "warning")
+
+    seq_len_calls <- source_expression$xml_find_function_calls("seq_len")
+    seq_len_expr <- xml_find_all(seq_len_calls, seq_len_xpath)
+    seq_len_lints <- xml_nodes_to_lints(
+      seq_len_expr,
+      source_expression,
+      "Use seq_along(x) instead of seq_len(length(x)).",
+      type = "warning"
+    )
+
+    c(seq_lints, seq_len_lints)
   })
 }
