### 2025-02-18 50df627
vectorized test
