+++ NEWS.md
@@ -33,6 +33,7 @@
 ## Notes
 
 * `expect_lint_free()` and other functions that rely on the {testthat} framework now have a consistent error message. (#2585, @F-Noelle).
+* `unnecessary_nesting_linter()` gives a more specific lint message identifying which "exit call" is unmatched & thus leading to the recommendation to reduce nesting (#2316, @MichaelChirico).
 
 # lintr 3.2.0
 
