### 2025-02-18 e20523d
Add a test of a custom exit call
