### 2025-02-18 833725a
recognize functional calls to undesirable operators
