+++ tests/testthat/test-sprintf_linter.R
@@ -4,14 +4,14 @@ patrick::with_parameters_test_that(
     linter <- sprintf_linter()
 
     # NB: using paste0, not sprintf, to avoid escaping '%d' in sprint fmt=
-    expect_no_lint(paste0(call_name, "('hello')"), linter)
-    expect_no_lint(paste0(call_name, "('hello %d', 1)"), linter)
-    expect_no_lint(paste0(call_name, "('hello %d', x)"), linter)
-    expect_no_lint(paste0(call_name, "('hello %d', x + 1)"), linter)
-    expect_no_lint(paste0(call_name, "('hello %d', f(x))"), linter)
-    expect_no_lint(paste0(call_name, "('hello %1$s %1$s', x)"), linter)
-    expect_no_lint(paste0(call_name, "('hello %1$s %1$s %2$d', x, y)"), linter)
-    expect_no_lint(paste0(call_name, "('hello %1$s %1$s %2$d %3$s', x, y, 1.5)"), linter)
+    expect_lint(paste0(call_name, "('hello')"), NULL, linter)
+    expect_lint(paste0(call_name, "('hello %d', 1)"), NULL, linter)
+    expect_lint(paste0(call_name, "('hello %d', x)"), NULL, linter)
+    expect_lint(paste0(call_name, "('hello %d', x + 1)"), NULL, linter)
+    expect_lint(paste0(call_name, "('hello %d', f(x))"), NULL, linter)
+    expect_lint(paste0(call_name, "('hello %1$s %1$s', x)"), NULL, linter)
+    expect_lint(paste0(call_name, "('hello %1$s %1$s %2$d', x, y)"), NULL, linter)
+    expect_lint(paste0(call_name, "('hello %1$s %1$s %2$d %3$s', x, y, 1.5)"), NULL, linter)
   },
   .test_name = c("sprintf", "gettextf"),
   call_name = c("sprintf", "gettextf")
@@ -57,9 +57,6 @@ patrick::with_parameters_test_that(
       rex::rex("invalid format '%d'; use format %f, %e, %g or %a for numeric objects"),
       linter
     )
-
-    # no arguments provided
-    expect_lint(paste0(call_name, "('hello')"), rex::rex("Omit sprintf()"), linter)
   },
   .test_name = c("sprintf", "gettextf"),
   call_name = c("sprintf", "gettextf")
@@ -69,23 +66,24 @@ test_that("edge cases are detected correctly", {
   linter <- sprintf_linter()
 
   # works with multi-line sprintf and comments
-  expect_no_lint(
+  expect_lint(
     trim_some("
       sprintf(
         'test fmt %s', # this is a comment
         2
       )
     "),
+    NULL,
     linter
   )
 
   # dots
-  expect_no_lint("sprintf('%d %d, %d', id, ...)", linter)
+  expect_lint("sprintf('%d %d, %d', id, ...)", NULL, linter)
 
   # TODO(#1265) extend ... detection to at least test for too many arguments.
 
   # named argument fmt
-  expect_no_lint("sprintf(x, fmt = 'hello %1$s %1$s')", linter)
+  expect_lint("sprintf(x, fmt = 'hello %1$s %1$s')", NULL, linter)
 
   expect_lint(
     "sprintf(x, fmt = 'hello %1$s %1$s %3$d', y)",
@@ -94,7 +92,7 @@ test_that("edge cases are detected correctly", {
   )
 
   # #2131: xml2lang stripped necessary whitespace
-  expect_no_lint("sprintf('%s', if (A) '' else y)", linter)
+  expect_lint("sprintf('%s', if (A) '' else y)", NULL, linter)
 })
 
 local({
[…]
