+++ tests/testthat/test-undesirable_operator_linter.R
@@ -3,8 +3,8 @@ test_that("linter returns correct linting", {
   msg_assign <- rex::rex("Avoid undesirable operator `<<-`.")
   msg_dollar <- rex::rex("Avoid undesirable operator `$`. As an alternative, use the `[[` accessor.")
 
-  expect_lint("x <- foo:::getObj()", NULL, linter)
-  expect_lint("cat(\"10$\")", NULL, linter)
+  expect_no_lint("x <- foo:::getObj()", linter)
+  expect_no_lint("cat(\"10$\")", linter)
   expect_lint(
     "a <<- log(10)",
     list(message = msg_assign, line_number = 1L, column_number = 3L),
@@ -23,15 +23,15 @@ test_that("undesirable_operator_linter handles '=' consistently", {
   linter <- undesirable_operator_linter(op = c("=" = "As an alternative, use '<-'"))
 
   expect_lint("a = 2L", rex::rex("Avoid undesirable operator `=`."), linter)
-  expect_lint("lm(data = mtcars)", NULL, linter)
-  expect_lint("function(a = 1) { }", NULL, linter)
+  expect_no_lint("lm(data = mtcars)", linter)
+  expect_no_lint("function(a = 1) { }", linter)
 })
 
 test_that("undesirable_operator_linter handles infixes correctly", {
   linter_oo <- undesirable_operator_linter(list("%oo%" = NA))
   linter_mod <- undesirable_operator_linter(list("%%" = NA))
   expect_lint("a %oo% b", rex::rex("Avoid undesirable operator `%oo%`."), linter_oo)
-  expect_lint("a %00% b", NULL, linter_oo)
+  expect_no_lint("a %00% b", linter_oo)
 
   # somewhat special case: %% is in infix_metadata
   expect_lint(
