### 2025-02-19 7f3039d
add non-linting examples (#2763)
Co-authored-by: AshesITR <alexander.rosenstock@web.de>
