### 2025-02-19 07c5584
Specific advice in unnecessary_nesting_linter metadata (#2759)
* attempt

* Add an early escape for get_r_string

* update tests

* vectorized test

* NEWS

* Add a test of a custom exit call

* better wording

Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>

---------

Co-authored-by: Indrajeet Patil <patilindrajeet.science@gmail.com>
