+++ R/unnecessary_nesting_linter.R
@@ -158,8 +158,7 @@ unnecessary_nesting_linter <- function(
       and not(IF)
       and not(expr[SYMBOL_FUNCTION_CALL[{xp_text_in_table(branch_exit_calls)}]])
     ]
-  ]
-  ")
+  ]")
   # condition for ELSE should be redundant, but include for robustness
   # condition on parent::expr[IF] ensures we're at the first `if` of a sequence of if/else statements
   # condition on expr uses following-sibling or preceding-sibling to ensure
@@ -168,9 +167,8 @@ unnecessary_nesting_linter <- function(
   #   to lead to a lint.
   # use position() = last() to ignore any expr but the last one in any branch.
   if_else_exit_xpath <- glue("
-  //expr[
-    IF
-    and ELSE
+  //IF/parent::expr[
+    ELSE
     and not(parent::expr[IF])
     and expr[
       OP-LEFT-BRACE
@@ -182,6 +180,8 @@ unnecessary_nesting_linter <- function(
     ]
   ]
   ")
+
+  used_exit_call_xpath <- glue("expr/expr[position() = last()]/{exit_call_expr}")
 
   assignment_cond <- if (allow_assignment) "expr[LEFT_ASSIGN or RIGHT_ASSIGN]" else "false"
 
@@ -245,14 +245,13 @@ unnecessary_nesting_linter <- function(
     xml <- source_expression$xml_parsed_content
 
     if_else_exit_expr <- xml_find_all(xml, if_else_exit_xpath)
+    used_exit_call <- get_r_string(if_else_exit_expr, used_exit_call_xpath)
     if_else_exit_lints <- xml_nodes_to_lints(
       if_else_exit_expr,
       source_expression = source_expression,
       lint_message = paste0(
         "Reduce the nesting of this if/else statement by unnesting the ",
-        "portion without an exit clause (i.e., ",
-        paste0(branch_exit_calls, "()", collapse = ", "),
-        ")."
+        "portion without an exit clause, i.e., ", used_exit_call, "()."
       ),
       type = "warning"
     )
