### 2025-02-19 d8f1f84
More hermetic tests
