### 2025-02-19 814f206
tests of globbing
