+++ tests/testthat/test-object_name_linter.R
@@ -40,21 +40,20 @@ test_that("styles are correctly identified", {
 
 test_that("linter ignores some objects", {
   # names for which style check is ignored
-  expect_lint("`%X%` <- t", NULL, object_name_linter("SNAKE_CASE")) # operator
-  expect_lint("`%x%` <- t", NULL, object_name_linter("snake_case")) # operator
-  expect_lint("`t.test` <- t", NULL, object_name_linter("UPPERCASE")) # std pkg
-  expect_lint(".Deprecated('x')", NULL, object_name_linter("lowercase")) # std pkg
-  expect_lint("print.foo <- t", NULL, object_name_linter("CamelCase")) # S3 generic
-  expect_lint("names.foo <- t", NULL, object_name_linter("CamelCase")) # int generic
-  expect_lint("sapply(x,f,USE.NAMES=T)", NULL, object_name_linter("snake_case")) # defined elsewhere
-  expect_lint(".onLoad <- function(...) TRUE", NULL, object_name_linter("snake_case")) # namespace hooks, #500
-  expect_lint(".First <- function(...) TRUE", NULL, object_name_linter("snake_case")) # namespace hooks
-  expect_lint("`%++%` <- `+`", NULL, object_name_linter("symbols")) # all-symbol operator
-  expect_lint("`%<-%` <- `+`", NULL, object_name_linter("symbols")) # all-symbol operator #495
+  expect_no_lint("`%X%` <- t", object_name_linter("SNAKE_CASE")) # operator
+  expect_no_lint("`%x%` <- t", object_name_linter("snake_case")) # operator
+  expect_no_lint("`t.test` <- t", object_name_linter("UPPERCASE")) # std pkg
+  expect_no_lint(".Deprecated('x')", object_name_linter("lowercase")) # std pkg
+  expect_no_lint("print.foo <- t", object_name_linter("CamelCase")) # S3 generic
+  expect_no_lint("names.foo <- t", object_name_linter("CamelCase")) # int generic
+  expect_no_lint("sapply(x,f,USE.NAMES=T)", object_name_linter("snake_case")) # defined elsewhere
+  expect_no_lint(".onLoad <- function(...) TRUE", object_name_linter("snake_case")) # namespace hooks, #500
+  expect_no_lint(".First <- function(...) TRUE", object_name_linter("snake_case")) # namespace hooks
+  expect_no_lint("`%++%` <- `+`", object_name_linter("symbols")) # all-symbol operator
+  expect_no_lint("`%<-%` <- `+`", object_name_linter("symbols")) # all-symbol operator #495
   # S3 group generic, #1841
-  expect_lint(
+  expect_no_lint(
     "`==.snake_case` <- function(a, b) unclass(a) == unclass(b)",
-    NULL,
     object_name_linter("snake_case")
   )
 })
@@ -63,8 +62,8 @@ test_that("linter returns correct linting", {
   lint_msg <- "Variable and function name style should match camelCase."
   linter <- object_name_linter("camelCase")
 
-  expect_lint("myObject <- 123", NULL, linter)
-  expect_lint("`myObject` <- 123", NULL, linter)
+  expect_no_lint("myObject <- 123", linter)
+  expect_no_lint("`myObject` <- 123", linter)
   expect_lint("my.confused_NAME <- 1;", list(message = lint_msg, line_number = 1L, column_number = 1L), linter)
   expect_lint("1 ->> read.data.frame;", list(message = lint_msg, line_number = 1L, column_number = 7L), linter)
   expect_lint(
@@ -81,14 +80,14 @@ test_that("linter returns correct linting", {
[…]
