### 2025-02-19 cdbf327
logic + tests + NEWS
