+++ tests/testthat/test-object_length_linter.R
@@ -2,7 +2,7 @@ test_that("returns the correct linting", {
   linter <- object_length_linter()
   lint_msg <- rex::rex("Variable and function names should not be longer than 30 characters.")
 
-  expect_lint("blah", NULL, linter)
+  expect_no_lint("blah", linter)
 
   expect_lint("very_very_very_very_long_variable_names_are_not_ideal <- 1", lint_msg, linter)
 
@@ -18,7 +18,7 @@ test_that("lints S3 generics correctly", {
   linter <- object_length_linter()
   lint_msg <- rex::rex("Variable and function names should not be longer than 30 characters.")
 
-  expect_lint("print.very_very_long_class_name <- 1", NULL, linter)
+  expect_no_lint("print.very_very_long_class_name <- 1", linter)
   expect_lint("print.very_very_very_very_long_class_name <- 1", lint_msg, linter)
 
   expect_lint(
@@ -78,7 +78,7 @@ test_that("function shorthand is caught", {
 test_that("rlang name injection is handled", {
   linter <- object_length_linter(length = 10L)
 
-  expect_lint("tibble('{foo() |> bar() |> baz()}' := TRUE)", NULL, linter)
+  expect_no_lint("tibble('{foo() |> bar() |> baz()}' := TRUE)", linter)
   expect_lint("DT[, 'a_very_long_name' := FALSE]", "names should not be longer than 10 characters", linter)
 })
 
