### 2025-02-20 cf31c59
NEWS terminology
