### 2025-02-20 6dba219
complete first pass of linter
