+++ R/null_coalescing_linter.R
@@ -44,28 +44,33 @@
 #' @seealso [linters] for a complete list of linters available in lintr.
 #' @export
 null_coalescing_linter <- function() {
-  xpath = "
+  braced_expr_cond <- "expr[1][OP-LEFT-BRACE and count(*) = 3]/expr"
+  xpath = glue("
   parent::expr[(
     preceding-sibling::IF
     and (
       expr[2] = following-sibling::ELSE/following-sibling::expr
+      or expr[2] = following-sibling::ELSE/following-sibling::{braced_expr_cond}
       or expr[2][LEFT_ASSIGN]/expr[1] = following-sibling::ELSE/following-sibling::expr
+      or expr[2][LEFT_ASSIGN]/expr[1] = following-sibling::ELSE/following-sibling::{braced_expr_cond}
     )
   ) or (
     preceding-sibling::OP-EXCLAMATION
     and parent::expr/preceding-sibling::IF
     and (
       expr[2] = parent::expr/following-sibling::expr[1]
+      or expr[2] = parent::expr/following-sibling::{braced_expr_cond}
       or expr[2][LEFT_ASSIGN]/expr[1] = parent::expr/following-sibling::expr[1]
+      or expr[2][LEFT_ASSIGN]/expr[1] = parent::expr/following-sibling::{braced_expr_cond}
     )
-  )]"
+  )]")
 
   Linter(linter_level = "expression", function(source_expression) {
-    xml <- source_expression$xml_parsed_content
+    null_calls <- source_expression$xml_find_function_calls("is.null")
 
-    bad_expr <- xml_find_all(xml, xpath)
+    bad_expr <- xml_find_all(null_calls, xpath)
 
-    is_negation <- !is.na(xml_find_first(bad_expr, "preceding-sibling::OP-EXCLAMATION")
+    is_negation <- !is.na(xml_find_first(bad_expr, "preceding-sibling::OP-EXCLAMATION"))
 
     observed <- ifelse(is_negation, "if (!is.null(x)) x else y", "if (is.null(x)) y else x")
 
