+++ man/best_practices_linters.Rd
@@ -48,6 +48,7 @@ The following linters are tagged with 'best_practices':
 \item{\code{\link{literal_coercion_linter}}}
 \item{\code{\link{nonportable_path_linter}}}
 \item{\code{\link{nrow_subset_linter}}}
+\item{\code{\link{null_coalescing_linter}}}
 \item{\code{\link{nzchar_linter}}}
 \item{\code{\link{object_overwrite_linter}}}
 \item{\code{\link{outer_negation_linter}}}
