+++ man/linters.Rd
@@ -17,17 +17,17 @@ see also \code{\link[=available_tags]{available_tags()}}.
 \section{Tags}{
 The following tags exist:
 \itemize{
-\item{\link[=best_practices_linters]{best_practices} (63 linters)}
+\item{\link[=best_practices_linters]{best_practices} (64 linters)}
 \item{\link[=common_mistakes_linters]{common_mistakes} (11 linters)}
 \item{\link[=configurable_linters]{configurable} (44 linters)}
-\item{\link[=consistency_linters]{consistency} (32 linters)}
+\item{\link[=consistency_linters]{consistency} (33 linters)}
 \item{\link[=correctness_linters]{correctness} (7 linters)}
 \item{\link[=default_linters]{default} (25 linters)}
 \item{\link[=efficiency_linters]{efficiency} (29 linters)}
 \item{\link[=executing_linters]{executing} (6 linters)}
 \item{\link[=package_development_linters]{package_development} (14 linters)}
 \item{\link[=pkg_testthat_linters]{pkg_testthat} (12 linters)}
-\item{\link[=readability_linters]{readability} (64 linters)}
+\item{\link[=readability_linters]{readability} (65 linters)}
 \item{\link[=regex_linters]{regex} (4 linters)}
 \item{\link[=robustness_linters]{robustness} (17 linters)}
 \item{\link[=style_linters]{style} (40 linters)}
@@ -97,6 +97,7 @@ The following linters exist:
 \item{\code{\link{nested_pipe_linter}} (tags: configurable, consistency, readability)}
 \item{\code{\link{nonportable_path_linter}} (tags: best_practices, configurable, robustness)}
 \item{\code{\link{nrow_subset_linter}} (tags: best_practices, consistency, efficiency)}
+\item{\code{\link{null_coalescing_linter}} (tags: best_practices, consistency, readability)}
 \item{\code{\link{numeric_leading_zero_linter}} (tags: consistency, readability, style)}
 \item{\code{\link{nzchar_linter}} (tags: best_practices, consistency, efficiency)}
 \item{\code{\link{object_length_linter}} (tags: configurable, default, executing, readability, style)}
