### 2025-02-20 6ba1b16
land on full 'if' expression, add metadata tests
