### 2025-02-20 29b1544
don't need %|||% in non-package file
