### 2025-02-20 f403721
Lint calls to undesirable operators in prefix form (#2761)
* scaffolding

* add argument

* recognize functional calls to undesirable operators

* revert incidentally included changes

* another one

* expect_no_lint

* terminology: prefix notation

* NEWS terminology
