### 2025-02-20 4186734
fix metadata & add more tests
