### 2025-02-21 7221565
Reference 'expression', not 'function', in metadata
