+++ R/cyclocomp_linter.R
@@ -22,16 +22,16 @@
 #' @seealso [linters] for a complete list of linters available in lintr.
 #' @export
 cyclocomp_linter <- function(complexity_limit = 15L) {
-  Linter(linter_level = "expression", function(source_expression) {
-    # nocov start
-    if (!requireNamespace("cyclocomp", quietly = TRUE)) {
-      cli::cli_abort(c(
-        "Cyclocomp complexity is computed using {.fn cyclocomp::cyclocomp}.",
-        i = "Please install the needed {.pkg cyclocomp} package."
-      ))
-    }
-    # nocov end
+  # nocov start
+  if (!requireNamespace("cyclocomp", quietly = TRUE)) {
+    cli::cli_abort(c(
+      "Cyclocomp complexity is computed using {.fn cyclocomp::cyclocomp}.",
+      i = "Please install the needed {.pkg cyclocomp} package."
+    ))
+  }
+  # nocov end
 
+  Linter(linter_level = "expression", function(source_expression) {
     complexity <- try_silently(
       cyclocomp::cyclocomp(parse(text = source_expression$content))
     )
@@ -45,7 +45,7 @@ cyclocomp_linter <- function(complexity_limit = 15L) {
       column_number = source_expression[["column"]][1L],
       type = "style",
       message = sprintf(
-        "Reduce the cyclomatic complexity of this function from %d to at most %d.",
+        "Reduce the cyclomatic complexity of this expression from %d to at most %d.",
         complexity, complexity_limit
       ),
       ranges = list(rep(col1, 2L)),
