+++ tests/testthat/test-cyclocomp_linter.R
@@ -3,7 +3,7 @@ test_that("returns the correct linting", {
 
   cc_linter_1 <- cyclocomp_linter(1L)
   cc_linter_2 <- cyclocomp_linter(2L)
-  lint_msg <- rex::rex("Reduce the cyclomatic complexity of this function")
+  lint_msg <- rex::rex("Reduce the cyclomatic complexity of this expression")
 
   expect_lint("if (TRUE) 1 else 2", NULL, cc_linter_2)
   expect_lint("if (TRUE) 1 else 2", lint_msg, cc_linter_1)
@@ -42,8 +42,20 @@ test_that("returns the correct linting", {
   expect_lint(complex_lines, lint_msg, cc_linter_2)
   expect_lint(
     complex_lines,
-    list(rex::rex("Reduce the cyclomatic complexity of this function from 10 to at most 2."), line_number = 1L),
+    list(rex::rex("Reduce the cyclomatic complexity of this expression from 10 to at most 2."), line_number = 1L),
     cc_linter_2
   )
   expect_lint(complex_lines, NULL, cyclocomp_linter(10L))
+
+  # no function involved, still lints, hence use 'expression' in the message
+  expect_lint(
+    trim_some("
+      for (i in 1:10)
+        for (j in 1:10)
+          for (k in 1:10)
+            i*j*k
+    "),
+    lint_msg,
+    cyclocomp_linter(5L)
+  )
 })
