### 2025-02-21 f00bf9d
skip running code that fails absent cyclocomp
