### 2025-02-21 27a6b5f
Deploying to gh-pages from @ r-lib/lintr@1118f0cae9596bbb2331f2ffeab568d30b6d5cd0 🚀
