### 2025-02-21 3a595e7
revert all but the minimal vocabulary change
