+++ tests/testthat/test-indentation_linter.R
@@ -1,13 +1,12 @@
 test_that("indentation linter flags unindented expressions", {
   linter <- indentation_linter(indent = 2L)
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       lapply(1:10, function(i) {
         i %% 2
       })
     "),
-    NULL,
     linter
   )
 
@@ -47,13 +46,12 @@ test_that("indentation linter flags unindented expressions", {
     linter
   )
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       lapply(1:10, function(i) {
           i %% 2
       })
     "),
-    NULL,
     indentation_linter(indent = 4L)
   )
 
@@ -68,50 +66,46 @@ test_that("indentation linter flags unindented expressions", {
   )
 
   # ugly code, but still correctly indented
-  expect_lint(
+  expect_no_lint(
     trim_some("
       list(
            1,
            2)
     "),
-    NULL,
     linter
   )
 
   # comments do not trigger hanging indent rule
-  expect_lint(
+  expect_no_lint(
     trim_some("
       list( # comment
         ok
       )
     "),
-    NULL,
     linter
   )
 
   # comments do not suppress block indents (#1751)
-  expect_lint(
+  expect_no_lint(
     trim_some("
       a <- # comment
         42L
     "),
-    NULL,
     linter
   )
 
   # assignment triggers indent
-  expect_lint(
+  expect_no_lint(
     trim_some("
       a <-
         expr(
           42
         )
     "),
-    NULL,
     linter
   )
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (cond)
         code
@@ -129,14 +123,13 @@ test_that("indentation linter flags unindented expressions", {
         code
       }
     "),
-    NULL,
     linter
   )
 })
 
 test_that("indentation linter flags improper closing curly braces", {
   linter <- indentation_linter(indent = 2L)
-  expect_lint(
+  expect_no_lint(
     trim_some("
       lapply(1:10, function(i) {
         {
@@ -144,7 +137,6 @@ test_that("indentation linter flags improper closing curly braces", {
         }
       })
     "),
-    NULL,
     linter
   )
 
@@ -162,19 +154,18 @@ test_that("indentation linter flags improper closing curly braces", {
 
 test_that("function argument indentation works in tidyverse-style", {
   linter <- indentation_linter()
-  expect_lint(
+  expect_no_lint(
     trim_some("
       function(a = 1L,
                b = 2L) {
         a + b
       }
     "),
-    NULL,
     linter
   )
 
   # new style (#1754)
-  expect_lint(
+  expect_no_lint(
     trim_some("
       function(
           a = 1L,
@@ -182,7 +173,6 @@ test_that("function argument indentation works in tidyverse-style", {
         a + b
       }
     "),
-    NULL,
     linter
   )
 
@@ -224,7 +214,7 @@ test_that("function argument indentation works in tidyverse-style", {
     linter
   )
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       function(
         a = 1L,
@@ -233,19 +223,17 @@ test_that("function argument indentation works in tidyverse-style", {
         a + b
       }
     "),
-    NULL,
     linter
   )
 
   # anchor is correctly found with assignments as well
-  expect_lint(
+  expect_no_lint(
     trim_some("
       test <- function(a = 1L,
                        b = 2L) {
         a + b
       }
     "),
-    NULL,
     linter
   )
 
[…]
