### 2025-02-22 9ad0af2
fix errors of linting for loops
