+++ tests/testthat/test-return_linter.R
@@ -64,7 +64,7 @@ test_that("Lint return on end of lambda function", {
 test_that("Do not lint if/else statements (with return) on end of function", {
   linter <- return_linter(return_style = "explicit")
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       function() {
         if (x) {
@@ -76,7 +76,6 @@ test_that("Do not lint if/else statements (with return) on end of function", {
         }
       }
     "),
-    NULL,
     linter
   )
 })
@@ -143,7 +142,7 @@ test_that("Lint control statements (without return) on end of function", {
 })
 
 test_that("Do not lint stop on end of function", {
-  expect_lint(
+  expect_no_lint(
     trim_some("
       function() {
         # Test
@@ -151,29 +150,26 @@ test_that("Do not lint stop on end of function", {
         stop(1)
       }
     "),
-    NULL,
     return_linter(return_style = "explicit")
   )
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       function() {
         stop(1)
       }
     "),
-    NULL,
     return_linter()
   )
 })
 
 test_that("return_linter works in simple function", {
-  expect_lint(
+  expect_no_lint(
     trim_some("
       foo <- function(bar) {
         return(bar)
       }
     "),
-    NULL,
     return_linter(return_style = "explicit")
   )
 })
@@ -181,27 +177,26 @@ test_that("return_linter works in simple function", {
 test_that("return_linter works for using stop() instead of returning", {
   linter <- return_linter(return_style = "explicit")
 
-  expect_lint(
+  expect_no_lint(
     trim_some("
       foo <- function(bar) {
         stop('bad')
       }
     "),
-    NULL,
     linter
   )
 })
 
 test_that("return_linter ignores expressions that aren't functions", {
-  expect_lint("x + 1", NULL, return_linter(return_style = "explicit"))
+  expect_no_lint("x + 1", return_linter(return_style = "explicit"))
 })
 
 test_that("return_linter ignores anonymous/inline functions", {
-  expect_lint("lapply(rnorm(10), function(x) x + 1)", NULL, return_linter(return_style = "explicit"))
+  expect_no_lint("lapply(rnorm(10), function(x) x + 1)", return_linter(return_style = "explicit"))
 })
 
 test_that("return_linter ignores if statements outside of functions", {
-  expect_lint(
+  expect_no_lint(
     trim_some("
       if (TRUE) {
         TRUE
@@ -209,20 +204,18 @@ test_that("return_linter ignores if statements outside of functions", {
         FALSE
       }
     "),
-    NULL,
     return_linter(return_style = "explicit")
   )
 })
 
 test_that("return_linter passes on multi-line functions", {
-  expect_lint(
+  expect_no_lint(
     trim_some("
       foo <- function(x) {
         y <- x + 1
         return(y)
       }
     "),
-    NULL,
     return_linter(return_style = "explicit")
   )
 })
@@ -302,13 +295,12 @@ test_that("return_linter works regardless of braces in final if case", {
     rex::rex("All functions must have an explicit return()."),
     linter
   )
-  expect_lint(
+  expect_no_lint(
     trim_some("
       foo <- function() {
         if (TRUE) return(TRUE)
       }
     "),
-    NULL,
     linter
   )
 })
@@ -349,7 +341,7 @@ test_that("return_linter finds missing return in one branch of an if", {
 test_that("return_linter works in nested if statements", {
   linter <- return_linter(return_style = "explicit")
 
-  expect_lint(
[…]
