### 2025-02-22 0bf8264
fix description of allow_file_path= (#2771)
