### 2025-02-22 2f94a5e
Land on function body more robustly in return_linter() (#2779)
