### 2025-02-22 675b9cd
Support * (globs) in file/directory exclusions (#2764)
