### 2025-02-24 5db4168
emphasize relative path of paths in config
