### 2025-02-24 5d248f3
changes for cyclocomp to support checking availability up-front
