### 2025-02-24 16d38f3
add a regression test
