### 2025-02-24 a2ed7c9
Adjust to "no lints found" output (#2784)
