+++ R/utils.R
@@ -104,8 +104,8 @@ get_content <- function(lines, info) {
   paste(lines, collapse = "\n")
 }
 
-logical_env <- function(x) {
-  res <- as.logical(Sys.getenv(x))
+logical_env <- function(x, unset = "") {
+  res <- as.logical(Sys.getenv(x, unset = unset))
   if (is.na(res)) {
     return(NULL)
   }
