### 2025-02-24 9adb04e
remove unused rot()
