### 2025-02-24 469f5d6
remove GITHUB_TOKEN support
