+++ man/default_settings.Rd
@@ -10,7 +10,7 @@
 \alias{.lintr}
 \title{Default lintr settings}
 \format{
-An object of class \code{list} of length 12.
+An object of class \code{list} of length 11.
 }
 \usage{
 default_settings
@@ -25,7 +25,6 @@ The default settings consist of
 \item \code{exclude_linter}, \code{exclude_linter_sep}: patterns used to exclude linters
 \item \code{exclusions}: a list of exclusions, see \code{\link[=exclude]{exclude()}} for a complete description of valid values.
 \item \code{cache_directory}: location of cache directory
-\item \code{comment_token}: a GitHub token character
 \item \code{error_on_lint}: decides if error should be produced when any lints are found
 }
 
