### 2025-02-24 c8db03b
switch to being an error to use the disabled linter
