### 2025-02-24 e62ec67
More generality for kwargs
