+++ tests/testthat/test-object_length_linter.R
@@ -95,4 +95,11 @@ test_that("literals in assign() and setGeneric() are checked", {
   expect_no_lint("setGeneric('good_name', function(x) x, package = 'badBadBadBadName')", linter)
   expect_lint("setGeneric('badBadBadBadName', function(x) x)", lint_msg, linter)
   expect_lint('setGeneric("badBadBadBadName", function(x) x)', lint_msg, linter)
+
+  # Ditto for keyword arguments
+  expect_no_lint("assign(envir = 'badBadBadBadName', nm, 2)", linter)
+  expect_no_lint("assign(envir = 'badBadBadBadName', 'good_name', 2)", linter)
+  expect_no_lint("assign(envir = 'badBadBadBadName', x = 'good_name', 2)", linter)
+  expect_lint("assign(envir = 'good_env_name', 'badBadBadBadName', 2)", lint_msg, linter)
+  expect_lint("assign(envir = 'good_env_name', x = 'badBadBadBadName', 2)", lint_msg, linter)
 })
