+++ tests/testthat/test-cyclocomp_linter.R
@@ -69,5 +69,5 @@ test_that("a null linter is returned, with warning, if cyclocomp is unavailable"
     linter <- cyclocomp_linter(1L)
   })
 
-  expect_error(lint(text = "if (TRUE) 1 else 2"), linters = linter), "disabled", fixed = TRUE)
+  expect_error(lint(text = "if (TRUE) 1 else 2", linters = linter), "disabled", fixed = TRUE)
 })
