### 2025-02-24 709480e
teach about .env pronoun
