### 2025-02-24 0bf5248
add a test of packages=
