+++ man/exclude.Rd
@@ -44,8 +44,8 @@ a character vector of files to exclude or a vector of lines to exclude.
 Note also that:
 \enumerate{
 \item All paths are interpreted as globs (\code{\link[=Sys.glob]{Sys.glob()}}), so that e.g. \code{*} does pattern expansion.
-\item For \code{exclusions} provided \emph{via} config, all paths are taken as relative \emph{to the config file itself},
-not the path where \code{lint()} is run.
+\item For \code{exclusions} provided \emph{via} a config file, all paths are taken as
+relative \emph{to the config file itself}, not the path where \code{lint()} is run.
 }
 }
 }
