+++ tests/testthat/test-undesirable_function_linter.R
@@ -3,8 +3,8 @@ test_that("linter returns correct linting", {
   msg_return <- rex::rex('Avoid undesirable function "return".', end)
   msg_log10 <- rex::rex('Avoid undesirable function "log10". As an alternative, use log().')
 
-  expect_lint("x <- options()", NULL, linter)
-  expect_lint("cat(\"Try to return\")", NULL, linter)
+  expect_no_lint("x <- options()", linter)
+  expect_no_lint("cat(\"Try to return\")", linter)
   expect_lint("lapply(x, log10)", list(message = msg_log10, line_number = 1L, column_number = 11L), linter)
   expect_lint("return()", list(message = msg_return, line_number = 1L, column_number = 1L), linter)
   expect_lint(
@@ -21,8 +21,8 @@ test_that("linter returns correct linting", {
     linter
   )
   # regression test for #1050
-  expect_lint("df$return <- 1", NULL, linter)
-  expect_lint("df@return <- 1", NULL, linter)
+  expect_no_lint("df$return <- 1", linter)
+  expect_no_lint("df@return <- 1", linter)
 })
 
 test_that("it's possible to NOT lint symbols", {
@@ -30,22 +30,22 @@ test_that("it's possible to NOT lint symbols", {
     fun = c(dir = NA, log10 = "use log()"),
     symbol_is_undesirable = FALSE
   )
-  expect_lint("dir <- 'path/to/a/directory'", NULL, linter)
-  expect_lint("lapply(x, log10)", NULL, linter)
+  expect_no_lint("dir <- 'path/to/a/directory'", linter)
+  expect_no_lint("lapply(x, log10)", linter)
 })
 
 test_that("undesirable_function_linter doesn't lint library and require calls", {
   linter <- undesirable_function_linter(fun = c(foo = NA))
   expect_lint("test::foo()", "undesirable", linter)
-  expect_lint("foo::test()", NULL, linter)
-  expect_lint("library(foo)", NULL, linter)
-  expect_lint("require(foo)", NULL, linter)
+  expect_no_lint("foo::test()", linter)
+  expect_no_lint("library(foo)", linter)
+  expect_no_lint("require(foo)", linter)
 
   linter <- undesirable_function_linter(fun = c(foo = NA, bar = NA))
-  expect_lint("library(foo)", NULL, linter)
+  expect_no_lint("library(foo)", linter)
 
   linter <- undesirable_function_linter(fun = c(foo = NA, bar = NA), symbol_is_undesirable = FALSE)
-  expect_lint("library(foo)", NULL, linter)
+  expect_no_lint("library(foo)", linter)
 })
 
 # regression test for #866
@@ -55,27 +55,19 @@ test_that("Line numbers are extracted correctly", {
 })
 
 test_that("invalid inputs fail correctly", {
-  error_msg <- "`fun` should be a non-empty named character vector"
-
-  expect_error(
-    undesirable_function_linter("***"),
-    error_msg,
-    fixed = TRUE
-  )
-  expect_error(
-    undesirable_function_linter(c("***" = NA, NA)),
-    error_msg,
-    fixed = TRUE
-  )
   expect_error(
     undesirable_function_linter(fun = NULL),
-    error_msg,
+    "`fun` should be a non-empty character vector",
     fixed = TRUE
   )
   expect_error(
     undesirable_function_linter(fun = character(0L)),
-    error_msg,
+    "`fun` should be a non-empty character vector",
     fixed = TRUE
+  )
+  expect_error(
+    undesirable_function_linter(c(NA, NA)),
+    "Found missing entries to `fun`: 1.*2"
   )
 
   expect_error(
[…]
