+++ man/undesirable_operator_linter.Rd
@@ -12,7 +12,7 @@ undesirable_operator_linter(
 \arguments{
 \item{op}{Named character vector. \code{names(op)} correspond to undesirable operators,
 while the values give a description of why the operator is undesirable.
-If \code{NA}, no additional information is given in the lint message. Defaults to
+If \code{NA} or unnamed, no additional information is given in the lint message. Defaults to
 \link{default_undesirable_operators}. To make small customizations to this list,
 use \code{\link[=modify_defaults]{modify_defaults()}}.}
 
