+++ R/undesirable_function_linter.R
@@ -38,7 +38,7 @@
 #'
 #' lint(
 #'   text = 'dir <- "path/to/a/directory"',
-#'   linters = undesirable_function_linter(fun = "dir"))
+#'   linters = undesirable_function_linter(fun = "dir")
 #' )
 #'
 #' # okay
@@ -59,7 +59,7 @@
 #'
 #' lint(
 #'   text = 'dir <- "path/to/a/directory"',
-#'   linters = undesirable_function_linter(fun = "dir", symbol_is_undesirable = FALSE))
+#'   linters = undesirable_function_linter(fun = "dir", symbol_is_undesirable = FALSE)
 #' )
 #'
 #' @evalRd rd_tags("undesirable_function_linter")
