+++ R/undesirable_operator_linter.R
@@ -31,6 +31,11 @@
 #'   linters = undesirable_operator_linter()
 #' )
 #'
+#' lint(
+#'   text = "mtcars$wt",
+#'   linters = undesirable_operator_linter("$")
+#' )
+#'
 #' # okay
 #' lint(
 #'   text = "a <- log(10)",
@@ -49,6 +54,11 @@
 #' lint(
 #'   text = "`:::`(utils, hasName)",
 #'   linters = undesirable_operator_linter(call_is_undesirable = FALSE)
+#' )
+#'
+#' lint(
+#'   text = 'mtcars[["wt"]]',
+#'   linters = undesirable_operator_linter("$")
 #' )
 #'
 #' @evalRd rd_tags("undesirable_operator_linter")
