### 2025-02-25 e38c518
add a test of available_linters(packages=) (#2789)
* add a test of packages=

* Add placeholder to namespace to allow mocking

* delint

* another test for coverage

* delint

---------

Co-authored-by: AshesITR <alexander.rosenstock@web.de>
