### 2025-02-25 9306f7c
Deploying to gh-pages from @ r-lib/lintr@e38c518af2e1a292b8fff7c57defc471c63b9ed6 🚀
