### 2025-02-25 84959ac
Emphasize that config paths are relative to the config itself (#2781)
* emphasize relative path of paths in config

* new attempt at wording more correctly

* delint

* better wording

Co-authored-by: AshesITR <alexander.rosenstock@web.de>

* deline & document

---------

Co-authored-by: AshesITR <alexander.rosenstock@web.de>
