### 2025-02-25 3a1e48b
Catch bad naming style (and length) in assign() + setGeneric() usages (#2766)
