### 2025-02-25 518983c
capture parser warnings as lints
