+++ tests/testthat/test-get_source_expressions.R
@@ -9,6 +9,7 @@ with_content_to_parse <- function(content, code) {
   content_env <- new.env()
   content_env$pc <- lapply(source_expressions[["expressions"]], `[[`, "parsed_content")
   content_env$error <- source_expressions$error
+  content_env$warning <- source_expressions$warning
   eval(substitute(code), envir = content_env)
 }
 
@@ -211,7 +212,7 @@ test_that("returned data structure is complete", {
   attr(lines_with_attr, "terminal_newline") <- TRUE
 
   exprs <- get_source_expressions(temp_file)
-  expect_named(exprs, c("expressions", "error", "lines"))
+  expect_named(exprs, c("expressions", "error", "warning", "lines"))
   expect_length(exprs$expressions, length(lines) + 1L)
 
   for (i in seq_along(lines)) {
@@ -448,6 +449,73 @@ test_that("Disallowed embedded null gives parser failure lint", {
   expect_lint(
     "'\\0'",
     rex::rex("Nul character not allowed."),
+    linters = list()
+  )
+})
+
+test_that("parser warnings are captured in output", {
+  with_content_to_parse("1e-3L", {
+    expect_length(warning, 1L)
+    expect_s3_class(warning, "lints")
+  })
+  with_content_to_parse("1e-3L; 1e-3L", {
+    expect_length(warning, 2L)
+  })
+  with_content_to_parse("1e-3L; 1.0L; 1.1L", {
+    expect_length(warning, 3L)
+  })
+  with_content_to_parse("1e-3L\n1.0L\n1.1L", {
+    expect_length(warning, 3L)
+  })
+  with_content_to_parse("1e-3L\n1+1\n1.0L\n2+2\n1.1L", {
+    expect_length(warning, 3L)
+  })
+})
+
+test_that("parser warnings generate lints", {
+  expect_lint(
+    "1e-3L",
+    "non-integer value 1e-3L",
+    linters = list()
+  )
+  expect_lint(
+    "1e-3L; 1e-3L",
+    list(
+      list("non-integer value 1e-3L", column_number = 1L),
+      list("non-integer value 1e-3L", column_number = 8L)
+    ),
+    linters = list()
+  )
+  expect_lint(
+    "1e-3L; 1.0L",
+    list(
+      list("non-integer value 1e-3L", column_number = 1L),
+      list("integer literal 1\\.0L", column_number = 8L)
+    ),
+    linters = list()
+  )
+  expect_lint(
+    trim_some("
+      1e-3L
+      1 + 1
+      1.0L
+      2 + 2
+      1.1L
+      3 + 3
+      2.2L
+      4 + 4
+      2.0L
+      5 + 5
+      2e-3L
+    "),
+    list(
+      list("non-integer value 1e-3L", line_number = 1L),
+      list("integer literal 1\\.0L", line_number = 3L),
+      list("integer literal 1.1L contains decimal", line_number = 5L),
+      list("integer literal 2\\.2L contains decimal", line_number = 7L),
+      list("integer literal 2\\.0L", line_number = 9L),
+      list("non-integer value 2e-3L", line_number = 11L)
+    ),
     linters = list()
   )
 })
