+++ R/get_source_expressions.R
@@ -51,6 +51,7 @@
 #'   }
 #'   }
 #'   \item{error}{A `Lint` object describing any parsing error.}
+#'   \item{warning}{A `lints` object describing any parsing warning.}
 #'   \item{lines}{The [readLines()] output for this file.}
 #' }
 #'
@@ -76,7 +77,7 @@ get_source_expressions <- function(filename, lines = NULL) {
   # Only regard explicit attribute terminal_newline=FALSE as FALSE and all other cases (e.g. NULL or TRUE) as TRUE.
   terminal_newline <- !isFALSE(attr(source_expression$lines, "terminal_newline", exact = TRUE))
 
-  e <- NULL
+  e <- w <- NULL
   source_expression$lines <- extract_r_source(
     filename = source_expression$filename,
     lines = source_expression$lines,
@@ -86,46 +87,55 @@ get_source_expressions <- function(filename, lines = NULL) {
   source_expression$content <- get_content(source_expression$lines)
   parsed_content <- get_source_expression(source_expression, error = function(e) lint_parse_error(e, source_expression))
 
-  if (is_lint(e) && (is.na(e$line) || !nzchar(e$line) || e$message == "unexpected end of input")) {
-    # Don't create expression list if it's unreliable (invalid encoding or unhandled parse error)
-    expressions <- list()
-  } else {
-    top_level_map <- generate_top_level_map(parsed_content)
-    xml_parsed_content <- safe_parse_to_xml(parsed_content)
-
-    expressions <- lapply(
-      X = top_level_expressions(parsed_content),
-      FUN = get_single_source_expression,
-      parsed_content,
-      source_expression,
-      filename,
-      top_level_map
-    )
-
-    if (!is.null(xml_parsed_content) && !is.na(xml_parsed_content)) {
-      expression_xmls <- lapply(
-        xml_find_all(xml_parsed_content, "/exprlist/*"),
-        function(top_level_expr) xml2::xml_add_parent(xml2::xml_new_root(top_level_expr), "exprlist")
-      )
-      for (i in seq_along(expressions)) {
-        expressions[[i]]$xml_parsed_content <- expression_xmls[[i]]
-        expressions[[i]]$xml_find_function_calls <- build_xml_find_function_calls(expression_xmls[[i]])
-      }
-    }
-
-    # add global expression
-    expressions[[length(expressions) + 1L]] <- list(
-      filename = filename,
-      file_lines = source_expression$lines,
-      content = source_expression$lines,
-      full_parsed_content = parsed_content,
-      full_xml_parsed_content = xml_parsed_content,
-      xml_find_function_calls = build_xml_find_function_calls(xml_parsed_content),
-      terminal_newline = terminal_newline
-    )
+  # Currently no way to distinguish the source of the warning
+  #   from the message itself, so we just grep the source for the
+  #   exact string generating the warning; de-dupe in case of
+  #   multiple exact matches like '1e-3L; 1e-3L'
+  if (length(w) > 0L) {
+    w <- unique(w)
+    w <- flatten_lints(lapply(w, lint_parse_warning, source_expression))
   }
 
-  list(expressions = expressions, error = e, lines = source_expression$lines)
+  if (is_lint(e) && (is.na(e$line) || !nzchar(e$line) || e$message == "unexpected end of input")) {
+    # Don't create expression list if it's unreliable (invalid encoding or unhandled parse error)
+    return(list(expressions = list(), error = e, warning = w, lines = source_expression$lines))
+  }
+
+  top_level_map <- generate_top_level_map(parsed_content)
+  xml_parsed_content <- safe_parse_to_xml(parsed_content)
[…]
