### 2025-02-25 b41cc97
helper to reduce cyclocomp
