### 2025-03-21 #7240 @IndrajeetPatil
@lrnv Nothing to update from my end; I don't have permissions to do any of this:

> @hadley I think that you need to archive the last version of lintr on Zenodo. For that, you need to connect with the r-lib account on zenodo IMHO, and once you are there there is an automatic archival tool
