### 2025-03-21 #7240 @lrnv
@IndrajeetPatil You need ot find out who can do it, and hadley already said he cannot. Maybe ask other lintr's developpers ?
