### 2025-03-21 #7240 @MichaelChirico
@hadley I think you need to give the r-lib organization Oauth authorization to Zenodo, I don't have the right permissions (no 'Settings' tab here) so it doesn't show up in my Zenodo dashboard.

![URL: https://github.com/r-lib](https://github.com/user-attachments/assets/466dbce4-97da-4828-bc9c-89cd747c18c0)

https://github.com/r-lib

Thanks in advance!
