### 2025-03-24 #7240 @hadley
But how? I don't see zenodo listed here.

<img width="1433" alt="Image" src="https://github.com/user-attachments/assets/988a270e-e2c2-4d4b-bf2b-
58f69fb776b5" />

 In the past, someone else has usually requested access and then I've approved it.
