### 2025-03-24 #7240 @MichaelChirico
Sorry I'm also lost. I briefly saw a screen when trying to log in to Zenodo to issue that request (IIRC) but I clicked my name instead and now I can't figure out how to navigate back to that.

I have this settings page on Zenodo, which lists authorized repos, but I don't see how to issue any new requests:

https://zenodo.org/account/settings/github/

Paging @lrnv for UX help here...
