### 2025-03-24 #7240 @MichaelChirico
Oh, I think I got it @hadley! You should have a request in your inbox.

I had to (1) disconnect GitHub from my settings page and then (2) Hit Connect again, which took me to the GitHub page where I could select which account/org to connect.
