### 2025-03-24 #7240 @MichaelChirico
Do I need to create a new release for this? I see `r-lib/lintr` listed here:

https://zenodo.org/account/settings/github/repository/r-lib/lintr

And we have an existing release I'd like to tag:

https://github.com/r-lib/lintr/releases/tag/v3.2.0

But I don't see it showing up on the Zenodo page.

(thanks everyone for the patience here)
