### 2025-03-24 #7240 @lrnv
@MichaelChirico What I need is a DOI from zenodo archiving the 3.2 version, so yes i think you need to create one. Then the current checklist for you was here : https://github.com/openjournals/joss-reviews/issues/7240#issuecomment-2655680714
