### 2025-03-25 #7240 @MichaelChirico
I'd really rather not create any duplicate releases just for this purpose, but I don't see any way to proceed. I filed https://github.com/zenodo/zenodo-rdm/issues/1139.
