### 2025-03-26 #7240 @lrnv
@MichaelChirico I think I already saw this issue myself... Ha there it is, a few years ago there was a way to trick zenodo into thinking that an old release just get made : https://github.com/zenodo/zenodo/issues/1463 Maybe you could try something like that ? Its a bit hacky, but...
