### 2025-03-26 #7240 @MichaelChirico
Excellent, thanks for the tip. I think maybe you need _any_ release in order to get a DOI assigned to the whole repo (?). Now https://doi.org/10.5281/zenodo.15090477 for {lintr} overall.

Anyway, here's the 3.2.0 DOI: https://zenodo.org/records/15090555

Thanks!
