### 2025-03-27 #7240 @lrnv
@MichaelChirico Can you please modify the archive of the 3.2 version so that the title and author list match the paper (currently there is no title and thousands of authors) ? Moreover, the licence should match the asoftware licence (currently CC by 4.0 on one side and MIT on the other)..
