### 2025-03-27 #7240 @MichaelChirico
Sorry, where are you seeing that? I looked at the 'payload' here to no avail https://zenodo.org/account/settings/github/repository/r-lib/lintr
