### 2025-03-31 #7240 @lrnv
Yes, its because the "trick" to release things from the previkous link does not release them in the right order / or rather its zenodo's fault, the releases are not ordered as they should. But this is not a problem for us. However:

* The title "r-lib/lintr: v3.2.0" on [zenodo.org/records/15090555](https://zenodo.org/records/15090555) should be the same as the JOSS paper's title...
