### 2025-03-31 #7240 @lrnv
@MichaelChirico Can you please go through the references and add DOI's to the ones that are missing (editorialbot just recomended a few) and check the "skipped" ones for potentially better references?
