### 2025-04-01 #7240 @MichaelChirico
I've edited the release's title as suggested.

@IndrajeetPatil would you mind following up on the DOIs issue? I confirmed that all of the suggested DOIs match those in the respective packages' `DOI:` key in their CRAN `DESCRIPTION`s.
