### 2025-04-02 #7240 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
