### 2025-04-03 #7240 @arfon
@jimhester et al – your paper is now accepted and published in JOSS :zap::rocket::boom:
