### 2025-05-14 #7240 @ritika-giri
Hello @jimhester  👋
This is impressive work! Have you considered submitting it to the [USRSE'25 ](https://us-rse.org/usrse25/participate/)conference ? I think it could be a great fit and would gain valuable visibility in the research software / open-science community. Submissions are due **May 16th**
Let me know if you'd be interested. I’d be happy to share more details!
